% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{abs_add_meta}
\alias{abs_add_meta}
\title{Add metadata to absorbance data}
\usage{
abs_add_meta(meta, abslist)
}
\arguments{
\item{meta}{a \code{data.frame} of metadata}

\item{abslist}{a \code{abslist} object containing the corresponding absorbance data}
}
\value{
a \code{abslist} object
}
\description{
Add metadata to absorbance data
}
\examples{
metadata <- meta_read(system.file("extdata", package = "eemanalyzeR"))
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern="ABS")
abs_agument <- abs_add_meta(metadata, abs)
}
