% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{abs_dir_read}
\alias{abs_dir_read}
\title{Read absorbance data files from directory}
\usage{
abs_dir_read(
  input_dir,
  pattern = NULL,
  skip = "SEM|BEM|Waterfall",
  file_ext = "dat",
  recursive = FALSE
)
}
\arguments{
\item{input_dir}{path to folder containing raw absorbance files}

\item{pattern}{optional. optional. a character string containing a \code{\link[base]{regular expression}} to be matched to the files in input_dir.
only files matching the pattern will be loaded.}

\item{skip}{optional. a character string containing a \code{\link[base]{regular expression}} to be matched to the files in input_dir.
any files matching this string will be ignored. useful for ignoring EEM's data.}

\item{file_ext}{character. the file extension of the EEMs}

\item{recursive}{logical. should the function recurse into directories?}
}
\value{
An object of class \code{abslist} containing a list of \code{abs}.
For more details see \link[eemanalyzeR]{abs_read}
}
\description{
A flexible function that will read all absorbance files in a directory even when
it contains other files.
}
\examples{
#load samples by using pattern to identify absorbance samples
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "abs|ABS")

#load samples by using skip to exclude EEM's and other samples
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), skip = "SEM|BEM|waterfall")

}
