% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-interp.R
\name{abs_interp}
\alias{abs_interp}
\title{Interpolate absorbance data}
\usage{
abs_interp(abs, type = "linear")
}
\arguments{
\item{abs}{An \code{abs} or \code{abslist} object.}

\item{type}{The interpolation method, either "linear" for linear interpolation
using \code{\link[zoo:na.approx]{zoo::na.approx()}}, or "spline" for spline interpolation using
\code{\link[zoo:na.spline]{zoo::na.spline()}}.}
}
\value{
An \code{abs} or \code{abslist} object.
}
\description{
Interpolates absorbance data to 1 nm resolution so indices can be calculated
at wavelengths that were not explicitly measured.
}
\examples{
abslist_filled <- abs_interp(example_abs)
abs_filled <- abs_interp(example_abs[[1]])
}
