% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abs_names}
\alias{abs_names}
\title{Get the names of abs or abslist objects}
\usage{
abs_names(abs)
}
\arguments{
\item{abs}{an object of class \code{abs} or \code{abslist}}
}
\value{
A character vector containing the names of the absorbance data
}
\description{
Get the names of abs or abslist objects
}
\examples{
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"))
names <- abs_names(abs)

abs_files <- list.files(system.file("extdata", package = "eemanalyzeR"),
full.names=TRUE, pattern="ABS")
abs <- abs_read(abs_files[1])
name <- abs_names(abs)
}
