% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-read.R
\name{abs_read}
\alias{abs_read}
\title{Read a single absorbance file into R}
\usage{
abs_read(file)
}
\arguments{
\item{file}{File path to the absorbance file to load.}
}
\value{
An object of class \code{abs} containing:
\itemize{
\item \strong{file}: The filename of the absorbance data
\item \strong{sample}: The name of the sample's file without the file extension
\item \strong{n}: The number of wavelengths measured
\item \strong{data}: A matrix where the first column is wavelength (nm) and the second column is absorbance
\item \strong{location}: Directory where the absorbance file is stored
}
}
\description{
Will accept absorbance files run via Aqualog exported using the sample Q or manually.
If the function tries to load an EEM file, it will return a warning and a \code{NULL} value for absorbance.
}
\examples{
abs_files <- list.files(
  system.file("extdata", package = "eemanalyzeR"),
  full.names = TRUE,
  pattern = "ABS|Abs ?Spectra ?Graphs")

abs <- abs_read(abs_files[1])
}
