% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_meta}
\alias{add_meta}
\alias{abs_add_meta}
\alias{eem_add_meta}
\title{Add metadata to absorbance and EEM's data}
\usage{
abs_add_meta(meta, abslist)

eem_add_meta(meta, eemlist)
}
\arguments{
\item{meta}{a \code{data.frame} of metadata}

\item{abslist}{an \code{abslist} object containing the corresponding absorbance data}

\item{eemlist}{an \code{eemlist} object containing the corresponding EEM's data}
}
\value{
returns an object of the same class as the object input into the function (\code{abslist} or \code{eemlist}) with metadata added in
The following items are added for each sample, if available in the metadata:
\itemize{
 \item file_name: the full data file name of the sample.
 \item dilution: the dilution factor for the sample.
 \item analysis_date: the date the sample was run.
 \item description: optional description of sample.
 \item doc_mgL: the concentration of dissolved organic carbon in the sample given in mg \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}
 \item notes: optional notes related to the sample or sample collection.
}
}
\description{
Add metadata to absorbance and EEM's data
}
\note{
if \code{eemlist} contains blanks, the blanks will get the metadata of the corresponding sample.
In the original \link[eemR]{eem} object, 'sample' is the file name, in the output from this function, 'sample' is replaced with the
'data_identifier' from the metadata, and the file name is moved to 'file_name'
}
\examples{

#add metadata to absorbance data
abs_augment <- abs_add_meta(metadata, example_absorbance)

#add metadata to EEM's data
eem_augment <- eem_add_meta(metadata, example_eems)
}
