% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to absorbance and EEM data}
\usage{
add_metadata(
  meta,
  x,
  iblank_pattern = "BEM$|Waterfall ?Plot ?Blank",
  sblank_pattern = "Blank|blank|BLK",
  check_pattern = "Tea|tea"
)
}
\arguments{
\item{meta}{A \code{data.frame} containing metadata.}

\item{x}{An \code{eemlist} or \code{abslist} object.}

\item{iblank_pattern}{a character vector of length 1 with a regular expression that matches sample names of instrument blanks. Default is "BEM$|Waterfall ?Plot ?Blank"}

\item{sblank_pattern}{a character vector of length 1 with a regular expression that matches sample names of sample blanks. Default is "Blank|blank|BLK"}

\item{check_pattern}{a character vector of length 1 with a regular expression that matches sample names of check standards. Default is "Tea|tea"}
}
\value{
An object of the same class as the input (\code{abslist} or \code{eemlist}), with metadata added.

For each sample, the following fields may be added (if present in the metadata):
\itemize{
\item \strong{sample_id}: sample identifier for the sample
\item \strong{dilution}: sample dilution factor
\item \strong{analysis_date}: date the sample was run
\item \strong{sample_name}: optional sample name or description
\item \strong{sample_type}: optional flag (e.g., \code{sample} for a sample,
\code{sblank} for an analytical blank, \code{check} for a check standard).
Default values match Horiba Aqualog exports.
\item \strong{doc_mgL}: dissolved organic carbon concentration in mg/L
\item \strong{notes}: optional notes on sample or sampling conditions
}

If \code{x} is an \code{eemlist}, two additional items are added:
\itemize{
\item \strong{integration_time_s}: integration time of the sample
\item \strong{raman_area_1s}: Raman water peak area normalized to 1-second integration time.
}
}
\description{
Adds metadata from a data frame to an \code{abslist} or \code{eemlist} object as additional
list items.
}
\note{
If an \code{eemlist} contains blanks, the blanks automatically inherit metadata from their
corresponding sample.
}
\examples{
# Add metadata to absorbance data
abs_augment <- add_metadata(metadata, example_abs)

# Add metadata to EEM data
eem_augment <- add_metadata(metadata, example_eems)
}
