% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{annotate_plot}
\alias{annotate_plot}
\title{Annotate EEMs plot with location of indices}
\usage{
annotate_plot(plot, index_method = "eemanalyzeR")
}
\arguments{
\item{plot}{a plot of an \code{eem}}

\item{index_method}{currently supports "eemanalyzeR", "eemR", and "usgs". See details for more information.}
}
\value{
an ggplot2 object
}
\description{
Adds labels showing the regions within an EEM that are used to calculate
the indices, customized to the specific index method used.
}
\examples{
plot <- plot(example_processed_eems[[3]])
annotate_plot(plot)
}
