% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-SNR.R
\name{calc_raman_SNR_sqrt}
\alias{calc_raman_SNR_sqrt}
\title{Calculate the Water Raman SNR ratio using Horiba's Square Root Method}
\usage{
calc_raman_SNR_sqrt(eem)
}
\arguments{
\item{eem}{EEM object to have signal to noise ratio calculated - must be a blank!}
}
\value{
numeric signal to noise ratio
}
\description{
The peak signal is measured at the water Raman peak intensity at 397 nm (for 350 nm
excitation) and the noise in a region where no Raman signal is present (450 nm).
For a perfect optical system there would be no signal at 450 nm since there is
no Raman emission there, however, all electro-optical systems have some levels
of stray light and noise, which will contribute to a signal at 450 nm. The above
formula assumes that the noise is governed by Poisson statistics and, therefore,
can be calculated as the square root of the baseline signal counts at 450 nm.
It is only applicable to photon counting detection, so for comparison purposes
it should only be used when comparing two photon counting spectrofluorometers.
}
