% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mdl.R
\name{check_eem_mdl}
\alias{check_eem_mdl}
\title{Check if a EEM sample is above the MDL}
\usage{
check_eem_mdl(eem, mdl = NULL, ex, em, vals = FALSE)
}
\arguments{
\item{eem}{an \code{eem} or \code{eemlist} object containing EEM's data}

\item{mdl}{a \code{eem} object containing MDL data, if no MLD provided will return NA}

\item{ex}{a vector of excitation wavelengths}

\item{em}{a vector of emission wavelengths}

\item{vals}{logical, if FALSE will return a TRUE or FALSE, if TRUE will
return a table with the eem and mdl values for the ex/em pair}
}
\value{
If \code{vals} is FALSE:
\itemize{
\item MDL01 if all values are below the MDL
\item MDL02 if some of the values are below the MDL
\item NA if all are above the MDL
}

If \code{vals} is TRUE:
\itemize{
\item a \code{data.frame} with four columns:
\itemize{
\item ex: the excitation wavelengths
\item em: the emission wavelengths
\item eem: the EEM fluorescence value for the ex/em pair
\item mdl: the MDL for the ex/em pair
}
}
}
\description{
Given a set of excitation and emission wavelengths, will check if the values in the
\code{eem} or \code{eemlist} are above the method detection limit (MDL)
for all excitation-emission pairs.
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_metadata(metadata, example_eems)
eemlist <- add_blanks(eemlist, validate=FALSE)
eemlist <- process_eem(eemlist, abslist)
mdl <- readRDS(file.path(system.file("extdata", package = "eemanalyzeR"),
"eem-mdl.rds"))

#works with a single sample
check_eem_mdl(eemlist[[1]], mdl, ex = 270:280, em=300:320)

#or an eemlist
check_eem_mdl(eemlist, mdl, ex = 270:280, em=300:320)
}
