% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mdl.R
\name{check_mdl}
\alias{check_mdl}
\alias{check_eem_mdl}
\alias{check_abs_mdl}
\title{Check whether an EEM or absorbance sample is above the MDL}
\usage{
check_eem_mdl(eem, mdl = NULL, ex, em, vals = FALSE)

check_abs_mdl(abs, mdl = NULL, wl, vals = FALSE)
}
\arguments{
\item{eem}{An \code{eem} or \code{eemlist} object.}

\item{mdl}{An \code{eem} or \code{abs} object containing MDL data.
If no MDL is provided, \code{NA} is returned.}

\item{ex}{A vector of excitation wavelengths.}

\item{em}{A vector of emission wavelengths.}

\item{vals}{If \code{TRUE}, returns actual values, otherwise returns just flags.}

\item{abs}{An \code{abs} or \code{abslist} object.}

\item{wl}{A vector of absorbance wavelengths.}
}
\value{
If \code{vals = FALSE}, returns one of:
\itemize{
\item "MDL01": all values are \strong{below} the MDL
\item "MDL02": some values are \strong{below} the MDL
\item \code{NA}: all values are \strong{above} the MDL
}

If \code{vals = TRUE}, returns:
\itemize{
\item A \code{data.frame} containing wavelengths, observed values, and MDL values
}
}
\description{
Checks whether values in an \code{eem} or \code{abs} object are above the method
detection limit (MDL) for a specified set of wavelengths or
excitation–emission pairs.
}
\examples{
# Load MDL data
eem_mdl <- readRDS(file.path(
  system.file("extdata", package = "eemanalyzeR"),
  "default-eem-mdl.rds"
))

abs_mdl <- readRDS(file.path(
  system.file("extdata", package = "eemanalyzeR"),
  "default-abs-mdl.rds"
))

# Single EEM sample
check_eem_mdl(example_processed_eems[[1]], eem_mdl,
              ex = 270:280, em = 300:320)

# EEM list
check_eem_mdl(example_processed_eems, eem_mdl,
              ex = 270:280, em = 300:320)

# Single absorbance sample
check_abs_mdl(example_processed_abs[[2]], abs_mdl, wl = 254)

# Absorbance list
check_abs_mdl(example_processed_abs, abs_mdl, wl = 254)
}
