% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_processing}
\alias{check_processing}
\title{Check which processing steps have been completed on an eem or eemlist}
\usage{
check_processing(eem)
}
\arguments{
\item{eem}{an \code{eemlist} or \code{eem} object containing EEM's data.}
}
\value{
a data.frame where the first column is the different attributes and the second is a T/F where F indicates that not all the samples
have been processed for that attribute.
}
\description{
Check which processing steps have been completed on an eem or eemlist
}
