% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-testing.R
\name{class-testing}
\alias{class-testing}
\alias{.is_eem}
\alias{.is_eemlist}
\alias{.is_abs}
\alias{.is_abslist}
\title{Checks if object is the Right class}
\source{
\code{.is_eem} and \code{.is_eemlist} functions were directly pulled from
\href{https://cran.r-project.org/web/packages/staRdom/index.html}{staRdom}.
}
\usage{
.is_eem(eem)

.is_eemlist(eem)

.is_abs(abs)

.is_abslist(abs)
}
\arguments{
\item{eem}{an object}

\item{abs}{an object}
}
\value{
a logical, \code{TRUE} or \code{FALSE} of if the object is that class
}
\description{
Utility functions to test class of objects, returning a logical.
}
