% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-testing.R
\name{class_testing}
\alias{class_testing}
\alias{.is_eem}
\alias{.is_eemlist}
\alias{.is_abs}
\alias{.is_abslist}
\title{Checks if object is the Right class}
\source{
\code{.is_eem} and \code{.is_eemlist} were directly pulled from
\href{https://cran.r-project.org/package=staRdom}{staRdom}.
}
\usage{
.is_eem(x)

.is_eemlist(x)

.is_abs(x)

.is_abslist(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) indicating whether the object
is of the expected class.
}
\description{
Utility functions to test class of objects, returning a logical.
}
\examples{
.is_eem(example_eems)
.is_eemlist(example_eems)

.is_abslist(example_abs)
.is_eem(example_abs)
}
