% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-dilution.R
\name{correct_dilution}
\alias{correct_dilution}
\title{Perform Dilution Corrections}
\usage{
correct_dilution(eemlist)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}
}
\value{
an object of class \code{eemlist}
}
\description{
Adjusts the EEM's data fluorescence to account for dilutions performed prior to measurement.
}
\details{
The function uses the dilution factor provided by the metadata to dilution correct the EEMs, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.
}
\examples{
eemlist <- add_metadata(metadata,example_eems)
eemlist <- eem_rm_blank(eemlist)
correct_eem <- correct_dilution(eemlist)
}
