% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-mdl.R
\name{create_mdl}
\alias{create_mdl}
\title{Calculate method detection limits (MDL)}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
\emph{Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance
and fluorescence from dissolved organic matter} (USGS Numbered Series No. 2018-1096).
U.S. Geological Survey. \doi{10.3133/ofr20181096}
}
\usage{
create_mdl(
  dir,
  method = "default",
  meta_file = NULL,
  sheet = NULL,
  iblank = "BEM",
  type = "eem",
  recursive = FALSE,
  qaqc_dir = get_qaqc_dir(),
  update_config = TRUE
)
}
\arguments{
\item{dir}{Path to a folder containing long-term EEMs and/or absorbance files.
All files in this directory will be loaded.}

\item{method}{A character string describing the method associated with the MDL files.}

\item{meta_file}{Name of the metadata file. Optional if the metadata file is the
only \code{.xlsx} or \code{.csv} file in \code{dir}. If not specified, the function attempts to find
a single metadata file and errors if multiple files are present.}

\item{sheet}{Name of the sheet containing metadata (only required
if the metadata is not on the first sheet of an \code{.xlsx} file).}

\item{iblank}{Optional. A character string containing a
\link[base:regex]{base::regular expression} used to identify instrument blanks.}

\item{type}{Which MDL to calculate: either "eem" or "abs".}

\item{recursive}{Logical. Should the function recursively search directories?}

\item{qaqc_dir}{Directory in which to save the QAQC \code{.rds} file.
Default: The QAQC directory in the current package environment, usually the user configured QAQC directory.
If \code{NA}, the function returns the MDL object instead of saving it.}

\item{update_config}{Logical. Should function ask to update user_config file with
the default QAQC directory location?}
}
\value{
\itemize{
\item If \code{qaqc_dir = FALSE}: returns an \code{eem} or \code{abs} object containing MDL values.
\item Otherwise: saves an \code{.rds} file containing the MDL object to the QAQC directory.
}
}
\description{
The method detection limit (MDL) is the minimum signal at which we can confidently
distinguish a measurement from zero and from analytical blanks. Here, the MDL is
calculated using the approach proposed by Hansen et al. (2018):
\strong{MDL = mean(long-term blank) + 3 × SD(long-term blank)}.
}
\details{
To calculate the MDL, you need:
\itemize{
\item A directory containing analytical blanks and instrument blanks
(fewer than 20 blanks triggers a warning).
\emph{Note: sample names must be unique.}
\item A metadata file for the blanks containing, at minimum, integration time
and Raman area, formatted as a metadata file described in \link{metadata}.
}
}
\examples{
eem_mdl <- create_mdl(
  file.path(system.file("extdata", package = "eemanalyzeR"), "long-term-blanks"),
  type = "eem",
  qaqc_dir = NA,
  update_config=FALSE
)

plot(eem_mdl)

}
