% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tea-std.R
\name{create_tea_std}
\alias{create_tea_std}
\title{Create Long Term Average Tea Standard}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence from dissolved organic matter
(USGS Numbered Series No. 2018-1096). Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence
from dissolved organic matter (Vol. 2018-1096). Reston, VA: U.S. Geological Survey. \url{https://doi.org/10.3133/ofr20181096}
}
\usage{
create_tea_std(
  dir,
  meta_name = NULL,
  sheet = NULL,
  eem_pattern = "Tea",
  abs_pattern = "Abs",
  blk = "BEM",
  type = "eem",
  recursive = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{dir}{path to folder containing long term EEMs and/or absorbance files}

\item{meta_name}{name of metadata file. optional if metadata is only xlsx or csv file in input_dir
if not specified function will attempt to load any xlsx or csv file in directory and return an error if there is more than one}

\item{sheet}{name of sheet containing metadata. only required if metadata isn't the first sheet}

\item{eem_pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the fluorescence data.}

\item{abs_pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the absorbance data.}

\item{blk}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the instrument blanks.}

\item{type}{which MDL to calculate: either \code{eem} or \code{abs}}

\item{recursive}{logical. should the function recurse into directories?}

\item{output_dir}{the location to save the mdl file to, default is a user-specific data directory (\link[rappdirs]{user_data_dir}). If
FALSE will return the averaged tea standard instead of saving.}
}
\value{
\itemize{
\item If output_dir is FALSE, returns an \code{eem} or \code{abs} object containing the average tea values.
\item Otherwise it saves a .RDS file containing the average tea data formatted as either an \code{eem} or \code{abs} object to the specified output_dir.
}
}
\description{
To ensure optical measurements are consistent across days, Hansen et al. (2018) recommend using a standard reference material:
Pure Leaf®, unsweetened black tea. The tea standard should be diluted to 1 \% concentration before analysis. It is recommended to
run the tea standard at the start and end of an analysis run. According to USGS standards, the indices for the tea standards run on
a given day should be within 20 \% of the long-term average (Hansen et al. 2018).
}
\details{
To calculate the average tea standard you need:
\itemize{
\item A directory containing tea standards with their associated instrument blanks (less than 20 will prompt a warning)
\itemize{
\item Note: sample names must be unique
}
\item Metadata for the tea samples including (at a minimum) the integration time and raman area in a single file,
formatted as a metadata file (see \link[eemanalyzeR]{metadata})
}
}
\examples{
eem_teastd <- create_tea_std(
file.path(system.file("extdata", package = "eemanalyzeR"), "long-term-tea"),
meta_name="longtermteastd-metadata.csv", eem_pattern = "BEM|SEM",
abs_pattern="ABS",type="eem", output_dir = FALSE)

plot(eem_teastd)

}
