% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_config}
\alias{default_config}
\title{Default configuration of arguments for eemanalyzeR}
\format{
A list of length 33:
\itemize{
\item \strong{abs_pattern}: Used by \code{\link[=abs_dir_read]{abs_dir_read()}}. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir}. Only files matching the pattern will be loaded.
\item \strong{abs_skip}: Used by \code{\link[=abs_dir_read]{abs_dir_read()}}. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir} that should be ignored.
\item \strong{abs_file_ext}: Used by \code{\link[=abs_dir_read]{abs_dir_read()}}. The file extension of the absorbance files.
\item \strong{abs_recurse_read}: Used by \code{\link[=eem_dir_read]{eem_dir_read()}}. Logical. Should the function recursively search directories?
\item \strong{eem_pattern}: Used by \code{\link[=eem_dir_read]{eem_dir_read()}}. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir}. Only files matching the pattern will be loaded.
\item \strong{eem_skip}: Used by \code{\link[=eem_dir_read]{eem_dir_read()}}. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir} that should be ignored.
\item \strong{eem_file_ext}: Used by \code{\link[=eem_dir_read]{eem_dir_read()}}. The file extension of the EEMs files.
\item \strong{eem_recurse_read}: Used by \code{\link[=eem_dir_read]{eem_dir_read()}}. Logical. Should the function recursively search directories?
\item \strong{eem_import_func}: Used by \code{\link[=abs_dir_read]{abs_dir_read()}}. Character or a user-defined function to import an EEM.
For more details, see \href{../doc/custom-indices.html}{\code{vignette("custom-indices")}}.
\item \strong{meta_file}: Used by \code{\link[=meta_read]{meta_read()}}. Name of the metadata file the program is expecting.
Defaults to NA, which auto-detects the metadata file in the data directory, for flexibility.
\item \strong{meta_sheet}: Used by \code{\link[=meta_read]{meta_read()}}. Name of the sheet containing metadata (only required if the
metadata is not on the first sheet of an \code{.xlsx} file).
\item \strong{meta_validate}: Used by \code{\link[=meta_read]{meta_read()}}. Logical. If \code{TRUE}, checks the metadata for
structural issues that could cause problems during processing.
Recommended to keep \code{TRUE}.
\item \strong{iblank_pattern}: Used by \code{\link[=add_metadata]{add_metadata()}}. A character vector of length 1 with a regular expression that
matches sample names of instrument blanks.
\item \strong{sblank_pattern}: Used by \code{\link[=add_metadata]{add_metadata()}}. A character vector of length 1 with a regular expression that
matches sample names of sample blanks.
\item \strong{check_pattern}: Used by \code{\link[=add_metadata]{add_metadata()}}. A character vector of length 1 with a regular expression that
matches sample names of check standards.
\item \strong{blank_validate}: Used by \code{\link[=add_blanks]{add_blanks()}}. Logical vector length one indicating whether blanks should be validated.
\item \strong{ex_clip}: Used by \code{\link[=process_eem]{process_eem()}}. Numeric vector of length two specifying the minimum and
maximum excitation wavelengths to keep.
\item \strong{em_clip}: Used by \code{\link[=process_eem]{process_eem()}}. Numeric vector of length two specifying the minimum and
maximum emission wavelengths to keep.
\item \strong{type}: Used by \code{\link[=process_eem]{process_eem()}}. Logical vector of length four indicating which scattering lines to remove.
The order is "raman1", "raman2", "rayleigh1", "rayleigh2".
\item \strong{width}: Used by \code{\link[=process_eem]{process_eem()}}. Numeric vector of length four specifying the width of scattering
lines to remove (nm). Same order as \code{type}.
\item \strong{interpolate}: Used by \code{\link[=process_eem]{process_eem()}}. Logical vector of length four indicating which scattering
lines to interpolate. Same order as \code{type}.
\item \strong{method}: Used by \code{\link[=process_eem]{process_eem()}}. Numeric (0–4) specifying the interpolation method to use.
Default is 1. See \code{\link[staRdom:eem_interp]{staRdom::eem_interp()}} for details.
\item \strong{cores}: Used by \code{\link[=process_eem]{process_eem()}}. Integer specifying the number of cores for parallel computation
during interpolation.
\item \strong{cuvle}: Used by \code{\link[=process_eem]{process_eem()}}. Cuvette (path) length in cm.
\item \strong{index_method}: Used by \code{\link[=get_indices]{get_indices()}}. Either "eemanalyzeR", "eemR", "usgs", or a custom
function.
\item \strong{tolerance}: Used by \code{\link[=get_indices]{get_indices()}}. Maximum percent deviation that the check standard can vary
from the long-term values without being flagged.
\item \strong{return}: Used by \code{\link[=get_indices]{get_indices()}}.Output format: "long" or "wide".
\item \strong{qaqc_dir}: Used by \code{\link[=get_indices]{get_indices()}}. File path to the QAQC files generated with \code{\link[=create_mdl]{create_mdl()}}
and \code{\link[=create_std]{create_std()}}. Default is a platform- and user-dependent data directory
\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}} created when QAQC files are generated.
\item \strong{qaqc_method}:
\item \strong{filename}: Used by \code{\link[=export_data]{export_data()}}. A character string, used to write output file names.
\item \strong{output_dir}: Used by \code{\link[=export_data]{export_data()}}. Path to save the data. Defaults to NA, which saves to a temporary directory
if not specified.
\item \strong{spectra_to_csv}: Used by \code{\link[=export_data]{export_data()}}. Logical. If \code{TRUE}, processed EEM and absorbance data and metadata are
written to \code{output_dir} as \code{.csv} files.
\item \strong{sum_plot}: Used by \code{\link[=export_data]{export_data()}}. Logical. If \code{TRUE}, individual plots are exported as a single large figure
along with the individual plots.
\item \strong{package_version}: the version of eemanalyzeR currently installed on the system.
}
}
\usage{
default_config
}
\description{
The default argument values used in the \code{\link[=run_eems]{run_eems()}} function. Argument
values are stored in a package environment (\code{.pkgenv}) which will be accessed
in the case where a specific variable is not specified by the user.
}
\details{
These values can be edited by:
\itemize{
\item the user permanently using \code{\link[=edit_user_config]{edit_user_config()}}
\item the user for the current R session using \code{modify_session_config()}
\item or within the \code{\link[=run_eems]{run_eems()}} function itself by providing argument values
}
}
\keyword{datasets}
