% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir-read.R
\name{dir_read}
\alias{dir_read}
\alias{eem_dir_read}
\alias{abs_dir_read}
\title{Read absorbance and fluorescence data from directory}
\usage{
eem_dir_read(
  input_dir,
  pattern = NA,
  skip = get_eem_skip(),
  file_ext = "dat",
  recursive = FALSE,
  import_function = "aqualog",
  verbose = TRUE
)

abs_dir_read(
  input_dir,
  pattern = NA,
  skip = get_abs_skip(),
  file_ext = "dat",
  recursive = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input_dir}{Path to folder containing raw EEMs and/or absorbance files.}

\item{pattern}{Optional. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir}. Only files matching the pattern will be loaded.}

\item{skip}{Optional. A character string containing a \code{\link[base:regex]{base::regular expression()}}
to match files in \code{input_dir} that should be ignored.}

\item{file_ext}{The file extension of the EEMs or absorbance files.}

\item{recursive}{Logical. Should the function recursively search directories?}

\item{import_function}{Character or a user-defined function to import an EEM.
For more details, see \href{../doc/custom-indices.html}{\code{vignette("custom-indices")}}.}

\item{verbose}{Logical. Should the function return notes about resetting the readme file?}
}
\value{
\itemize{
\item \code{eem_dir_read()} returns an object of class \code{eemlist}, containing a list of \code{eem}. For more details see \code{\link[eemR:eem_read]{eemR::eem_read()}}.
\item \code{abs_dir_read()} returns an object of class \code{abslist}, containing a list of \code{abs}. For more details see \code{\link[=abs_read]{abs_read()}}.
}
}
\description{
Wrappers for \code{\link[eemR:eem_read]{eemR::eem_read()}} and \code{\link[=abs_read]{abs_read()}} to read all EEMs or absorbance
files in a directory into R, even when the directory contains other files.
}
\examples{
# Load all EEMs from directory
eem_list <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"))

# Load all EEMs matching a pattern from directory
eem_list <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "SEM")

# Load absorbance samples using a pattern
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "abs|ABS")

# Load absorbance samples while skipping EEMs and other files
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), skip = "SEM|BEM|Waterfall")
}
