% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-index.R
\name{.combine_flags}
\alias{.combine_flags}
\title{Nicely combine data QAQC flags}
\usage{
.combine_flags(x, x1, mdl = FALSE)
}
\arguments{
\item{x}{existing flags}

\item{x1}{flags to add}

\item{mdl}{logical, combing two MDL flags?}
}
\description{
If the previous flags were `NA`, replaces with the new flagged value, otherwise combines
the flags with a "_" between.
}
\examples{
.combine_flags("DATA01", NA)
.combine_flags(NA, "MDL01")
.combine_flags(NA, NA)
.combine_flags("DATAO1", "MDL01")
.combine_flags("DATA01", "DATA01")
}
