% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{edit_config}
\alias{edit_config}
\title{Set up user defaults for data processing}
\usage{
edit_config()
}
\value{
Opens up `user-config.yaml`.
}
\description{
Opens up a YAML configuration file which stores the default arguments for
`eemanalyzeR`. This allows the user to change the package defaults by computer
customizing the processing steps in [run_eems] but also ensuring that all data
processed will use the same user defaults without needing to specify each time.
}
\details{
The defaults are stored in a YAML configuration file on the user data directory.
This function will open up the file so the text can be editted. To save new defaults simply
edit the file and save it. The arguments in this file will overwrite the defaults
set in the package.
}
\examples{
edit_config()
}
