% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{eem_add_meta}
\alias{eem_add_meta}
\title{Add metadata to EEM's data}
\usage{
eem_add_meta(meta, eemlist)
}
\arguments{
\item{meta}{a \code{data.frame} of metadata}

\item{eemlist}{a \code{eemlist} object containing the corresponding EEM's data}
}
\value{
a \code{eemlist} object
}
\description{
Add metadata to EEM's data
}
\examples{
eem_agument <- eem_add_meta(metadata, example_samples)
}
