% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{eem_dir_read}
\alias{eem_dir_read}
\title{Read excitation-emission fluorescence matrices (EEM) from directory}
\usage{
eem_dir_read(
  input_dir,
  pattern = NULL,
  skip = "(?i)abs",
  file_ext = "dat",
  recursive = FALSE,
  import_function = "aqualog"
)
}
\arguments{
\item{input_dir}{path to folder containing raw EEMs files}

\item{pattern}{optional. a character string containing a \code{\link[base]{regular expression}} to be matched to the files in input_dir.
only files matching the pattern will be loaded.}

\item{skip}{a character string containing a \code{\link[base]{regular expression}} to be matched to the files in input_dir.
any files matching this string will be ignored. useful for ignoring absorbance data.}

\item{file_ext}{character. the file extension of the EEMs}

\item{recursive}{logical. should the function recurse into directories?}

\item{import_function}{character or a user-defined function to import an EEM. for more details see \link[eemR]{eem_read}}
}
\value{
object of class \code{eemlist}. See \link[eemR]{eem_read} for more details on class structure.
}
\description{
A wrapper for \link[eemR]{eem_read} to read all EEMs in a directory even when
it contains other files.
}
\examples{
# Load all data from directory --------------
eemlist <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"))

# Load all data that matches a pattern from directory -------------
eemlist <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern="SEM")

}
