% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eem_get_info}
\alias{eem_get_info}
\title{Extract information from an eem or eemlist object}
\usage{
eem_get_info(eem, info)
}
\arguments{
\item{eem}{an object of class \code{eem} of \code{eemlist}}

\item{info}{the name of the component within the \code{eem} to extract.
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{add_meta}
for extended \code{eem} component names.}
}
\value{
a vector containing the info from the EEM's
}
\description{
This is a helper function that builds upon the \link[eemR]{eem_names} function by
extending it to extract any component of the \code{eem} as a vector.
}
\examples{
#get names
eem_get_info(example_eems, "sample")

#get analysis_date
eemlist <- eem_add_meta(meta, example_eems)
eem_get_info(eemlist, "analysis_date")

#get doc
eemlist <- eem_add_meta(meta, example_eems)
eem_get_info(eemlist, "doc_mgL")
}
