% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eem_normalize}
\alias{eem_normalize}
\title{Normalize an eem or eemlist based on a normalization factor}
\usage{
eem_normalize(eem, factor = NULL)
}
\arguments{
\item{eem}{the \code{eem} or \code{eemlist} to normalize}

\item{factor}{the normalization factor, either a single value or vector of factors, if NULL it will normalize to the maximum value for each eem}
}
\value{
an \code{eem} or \code{eemlist} where \code{x} has been normalized
}
\description{
Useful for raman normalization or normalizing to a max of one for blank comparisons.
}
\examples{
eem_normal <- eem_normalize(example_eems[1])
eems_normal <- eem_normalize(example_eems)

}
