% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eem_subtract}
\alias{eem_subtract}
\title{Subtract one eem from another}
\usage{
eem_subtract(eem1, eem2)
}
\arguments{
\item{eem1}{the eem that will be returned}

\item{eem2}{the eem that will be used for subtraction}
}
\value{
an \code{eem}
}
\description{
Useful for performing blank subtraction
}
\examples{
eem_sub <- eem_subtract(example_eems[[1]], longterm_blank)
}
