% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eem_unique}
\alias{eem_unique}
\title{Get unique EEM's}
\usage{
eem_unique(eemlist)
}
\arguments{
\item{eemlist}{an object of class \code{eemlist}}
}
\value{
an object of class \code{eemlist} with samples with duplicate EEM's matrices (eem$x) removed.

Note: sample names remain unchanged.
}
\description{
Get unique EEM's
}
\examples{
unique_eems <- eem_unique(example_blanks)
}
