% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemanalyzeR-package.R
\docType{package}
\name{eemanalyzeR-package}
\alias{eemanalyzeR}
\alias{eemanalyzeR-package}
\title{eemanalyzeR: Semi-Automated Processing of Excitation Emission Matrices}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for processing and analyzing raw excitation-emission matrices (EEMs) and absorbance data. Includes functions for data cleaning, quality assurance, calculation of indices, and generating visualizations. Builds upon the 'eemR' and 'staRdom' packages to create a more streamlined and automated workflow that ensures consistent analysis and simplify the coding required to process the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://katiewampler.github.io/eemanalyzeR/}
  \item \url{https://CRAN.R-project.org/package=eemanalyzeR},
  \item Report bugs at \url{https://github.com/katiewampler/eemanalyzeR/issues}
}

}
\author{
\strong{Maintainer}: Katie A. Wampler \email{katie.wampler@oregonstate.edu} (\href{https://orcid.org/0000-0002-5584-7554}{ORCID})

Authors:
\itemize{
  \item Katie A. Wampler \email{katie.wampler@oregonstate.edu} (\href{https://orcid.org/0000-0002-5584-7554}{ORCID})
  \item Ryan P. Cole \email{ryan.cole@oregonstate.edu}
}

}
\keyword{internal}
