% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_absorbance}
\alias{example_absorbance}
\title{Down-scaled absorbance data}
\format{
A data.frame with 32 rows and 4 columns:
\describe{
  \item{wavelength}{Wavelength absorbance was measured at in nanometers}
  \item{ExampleBlank}{Absorbance measured for a sample of Type 1 water}
  \item{ExampleTeaStd}{Absorbance measured for a sample of dilute unsweetened tea}
  \item{ExampleSample}{Absorbance measures for a sample of stream water from Oregon, USA}
}
}
\source{
Oregon State University Forest Watershed Hydrology Lab (2022-11-14)
}
\usage{
example_absorbance
}
\description{
A subset of water samples run on a Horiba Aqualog for absorbance. This data only includes
the absorbance data associated with the samples in \link{example_samples}.
}
\keyword{datasets}
