% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_processed_abs}
\alias{example_processed_abs}
\title{Down-scaled and processed absorbance data}
\format{
An \code{abslist} with four samples:
\itemize{
\item \strong{B1S1ExampleBlankABS}: Absorbance measured for a sample of Type 1 water
\item \strong{B1S2ExampleTeaStdABS}: Absorbance measured for a sample of dilute unsweetened tea
\item \strong{B1S3ExampleSampleABS}: Absorbance measured for a sample of stream water from Oregon, USA
\item \strong{ManualExampleTeaAbsSpectraGraphs}: Absorbance measured for a sample of dilute unsweetened tea
}
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022-11-14)
}
\usage{
example_processed_abs
}
\description{
Processed version of \code{\link[=example_abs]{example_abs()}} using \code{eemanalyzeR}.
}
\keyword{datasets}
