% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-data.R
\name{export_data}
\alias{export_data}
\title{Export processed absorbance and fluorescence data}
\usage{
export_data(
  eemlist,
  abslist,
  filename,
  meta = NULL,
  indices = NULL,
  eem_plot = NULL,
  abs_plot = NULL,
  csv = FALSE,
  sum_plot = TRUE,
  output_dir = NA
)
}
\arguments{
\item{eemlist}{An \code{eemlist} object.}

\item{abslist}{An \code{abslist} object.}

\item{filename}{A character string, used for file names.}

\item{meta}{A \code{data.frame} containing sample metadata.}

\item{indices}{If \code{NULL}, indices will not be exported. If a list of
indices is provided, it is saved.}

\item{eem_plot}{If \code{NULL}, EEM plots will not be exported. If a
list of plots is provided, they are saved as \code{.png} files. See \code{\link[=plot.eemlist]{plot.eemlist()}} for plotting.}

\item{abs_plot}{If \code{NULL}, absorbance plots will not be exported.
If a plot is provided, it is saved as a \code{.png} file. See \code{\link[=plot.abslist]{plot.abslist()}} for plotting.}

\item{csv}{Logical. If \code{TRUE}, processed EEM and absorbance data and metadata are
written to \code{output_dir} as \code{.csv} files.}

\item{sum_plot}{Logical. If \code{TRUE}, individual plots are exported as a single large figure
along with the individual plots.}

\item{output_dir}{Path to save the data. Defaults to a temporary directory
if not specified.}
}
\value{
A list of the processed data which is also saved to the output directory specified.
The list contains:
\itemize{
\item \strong{eemlist:} the \code{eemlist}
\item \strong{abslist:} the \code{abslist}
\item \strong{readme:} a character vector containing information about the data processing steps
\item \strong{meta:} the metadata associated with the samples, may be \code{NULL} if not provided
\item \strong{indices:} a list of EEMs and absorbance indices, may be \code{NULL} if not provided
\item \strong{eem_plot:} a list of EEMs plots, may be \code{NULL} if not provided
\item \strong{abs_plot:} a ggplot2 object of the absorbance data, may be \code{NULL} if not provided
}
}
\description{
Exports processed EEM and absorbance data, plots, indices, and metadata to a
specified output directory. Creates a folder named by \code{filename} and saves all specified outputs there.
}
\examples{
eem_plots <- plot(example_processed_eems)
abs_plot <- plot(example_processed_abs)

data <- export_data(
  eemlist = example_processed_eems,
  abslist = example_processed_abs,
  filename = "eemanalyzeR_example",
  eem_plot = eem_plots,
  abs_plot = abs_plot,
  meta = metadata
)
}
