% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract-blanks}
\alias{extract-blanks}
\alias{eem_get_blank}
\alias{eem_rm_blank}
\title{Extract blanks from eemlist}
\usage{
eem_get_blank(eemlist, pattern, info = "sample")

eem_rm_blank(eemlist, pattern, info = "sample")
}
\arguments{
\item{eemlist}{an object of class \code{eemlist}}

\item{pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the blanks.}

\item{info}{the name of the component within the \code{eem} to check for the pattern. default is 'sample'}
}
\value{
an object of class \code{eemlist} either with only the blanks (\code{eem_get_blank})
or only the samples (\code{eem_rm_blank})
}
\description{
Using a regular expression, the \code{eemlist} will be cut to include either
only the blank samples (\code{eem_get_blank}) or remove all blank samples (\code{eem_rm_blank}).
This function is similar to the \link[eemR]{eem_remove_blank} function, except it is more
flexible to include different patterns for the blank.
}
\note{
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{add_meta}
for extended \code{eem} component names.
}
\examples{
blanks <- eem_get_blank(example_eems, pattern = "BEM")
samples <- eem_rm_blank(example_eems, pattern = "BEM")
}
