% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tea.R
\name{abs_get_tea}
\alias{abs_get_tea}
\title{Extract tea from abslist}
\usage{
abs_get_tea(abslist, pattern = "tea", info = "sample")
}
\arguments{
\item{abslist}{an object of class \code{abslist}}

\item{pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the blanks.}

\item{info}{the name of the component within the \code{abs} to check for the pattern. default is 'sample'}
}
\value{
an object of class \code{abslist} with only the tea standards.
}
\description{
Using a regular expression, the \code{abslist} will be cut to include either
only the tea samples (\code{abs_get_tea}).
}
\note{
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{add_metadata}
for extended \code{eem} component names.
}
\examples{
tea <- abs_get_tea(example_abs, pattern = "tea")
}
