% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-index.R
\name{format_index}
\alias{format_index}
\title{Format index and flag data}
\usage{
format_index(x, index, value, flag)
}
\arguments{
\item{x}{An \code{eemlist} or \code{abslist} object.}

\item{index}{A character string giving the name of the index.}

\item{value}{Index values. Use \code{\link[=get_fluorescence]{get_fluorescence()}} or
\code{\link[=get_absorbance]{get_absorbance()}} to generate these values.}

\item{flag}{Flag values, the same length as \code{value}.}
}
\value{
A data frame with four columns:
\itemize{
\item \strong{sample_id}: sample identifier for the sample (if provided, otherwise uses sample: the sample's file name)
\item \strong{sample_name}: sample name or description (if provided, otherwise uses sample: the sample's file name)
\item \strong{index}: the name of the index
\item \strong{value}: the index value (with \code{MDL01} values removed)
}
}
\description{
Combines index values, QA/QC flags, and sample metadata into a clean
\code{data.frame} for further processing. Any values flagged as "MDL01"
(below detection limit) are removed and not reported.
}
\examples{
ex <- 240:260
em <- 300:320

vals <- get_fluorescence(example_eems, ex, em, stat = "max")
flags <- flag_missing(example_eems, ex = ex, em = em, all = FALSE)

index_formatted <- format_index(
  x = example_eems,
  index = "test_index",
  value = vals,
  flag = flags
)
}
