% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-absorbance.R
\name{get_absorbance}
\alias{get_absorbance}
\title{Extract absorbance at a given wavelength}
\usage{
get_absorbance(abs, wl, cuvle = 1, suva = FALSE)
}
\arguments{
\item{abs}{an object of class \code{abs} or \code{abslist}}

\item{wl}{the wavelength in (nm) to get the absorbance at}

\item{cuvle}{cuvette (path) length in cm}

\item{suva}{logical, if TRUE, returns the specific absorbance}
}
\value{
A vector of absorbance values. If a value cannot be extracted, NA will be returned.
}
\description{
Get the absorbance value with units of \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^-1}} at a specified wavelength with the option to
calculate specific absorbance (SUVA) which has units of \ifelse{html}{\out{m<sup>-1</sup>}}{\eqn{m^-1}}.
Note that metadata must be added in order for specific absorbance to be calculated using \link[eemanalyzeR]{add_metadata} or NA will be returned.
}
\details{
If the wavelength requested is not
in the absorbance data, it will be interpolated first using \link[eemanalyzeR]{abs_interp}.
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)

a254 <- get_absorbance(abslist, 254)
suva254 <- get_absorbance(abslist, 254, suva=TRUE)
}
