% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fluorescence.R
\name{get_fluorescence}
\alias{get_fluorescence}
\title{Get fluorescence within a specified range}
\usage{
get_fluorescence(eem, ex, em, stat = "max", norm = FALSE)
}
\arguments{
\item{eem}{An \code{eem} or \code{eemlist} object.}

\item{ex}{A vector of excitation wavelengths.}

\item{em}{A vector of emission wavelengths.}

\item{stat}{A string specifying the statistic to return:
\itemize{
\item "max": return the maximum value (default)
\item "sum": return the sum of values
}}

\item{norm}{Logical. If \code{TRUE}, divides the index value by the DOC
concentration. Metadata must be added using
\code{\link[=add_metadata]{add_metadata()}}, otherwise \code{NA} will be returned.}
}
\value{
A numeric vector of fluorescence values. Returns \code{NA} if values
cannot be extracted.
}
\description{
Gets fluorescence within a range of excitation and emission wavelengths and
returns either the maximum value or the sum across that range.
Missing values are interpolated using \code{\link[pracma:interp2]{pracma::interp2()}}.
}
\examples{
pA <- get_fluorescence(example_eems, ex = 250:260, em = 380:480)
pD <- get_fluorescence(example_eems, ex = 390, em = 509)

pA_sum <- get_fluorescence(
  example_eems,
  ex = 250:260,
  em = 380:480,
  stat = "sum"
)

eemlist <- add_metadata(metadata, example_eems)

pA_docnorm <- get_fluorescence(
  eemlist,
  ex = 250:260,
  em = 380:480,
  norm = TRUE
)
}
