% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indices.R
\name{get_indices_function}
\alias{get_indices_function}
\title{Get preset or use custom functions to generate fluorescence and absorbance indices}
\usage{
get_indices_function(method = "eemanalyzeR")
}
\arguments{
\item{method}{currently supports "eemanalyzeR", "eemR", and "usgs". See details for more information.}
}
\value{
a function used to generate indices
}
\description{
Get preset or use custom functions to generate fluorescence and absorbance indices
}
