% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_info}
\alias{get_info}
\title{Extract info from eemlist or abslist}
\usage{
get_info(x, info)
}
\arguments{
\item{x}{an object of class:
\itemize{
\item{\code{eemlist}}
\item{\code{eem}}
\item{\code{abslist}}
\item{\code{abs}}
}}

\item{info}{the name of the component within the \code{eem} or \code{abs} to extract.
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{abs_read} for base \code{abs} names.
see \link[eemanalyzeR]{add_meta} for extended \code{eem} component names.}
}
\value{
returns a vector containing the info from the EEM's or absorbance.
}
\description{
Helper function that build upon the \link[eemR]{eem_names}
function. Extracts components of an \code{eemlist/eem} and \code{abslist/abs} object.
}
\examples{
#get names
get_info(example_eems, "sample")
get_info(example_absorbance, "sample")

#get analysis_date
eemlist <- eem_add_meta(metadata, example_eems)
get_info(eemlist, "analysis_date")

#get doc for fifth eem
eemlist <- eem_add_meta(metadata, example_eems)
get_info(eemlist[[5]], "doc_mgL")
}
