% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ratios.R
\name{get_ratios}
\alias{get_ratios}
\title{Safely return ratios values}
\usage{
get_ratios(val1, val2)
}
\arguments{
\item{val1}{the numerator of the ratio, can be a single value or a vector of values}

\item{val2}{the denominator of the ratio, can be a single value or a vector of values}
}
\value{
a ratio value if possible or a character flag
}
\description{
Will either return a ratio value or a QA/QC flag if ratio value can't be calculated.
}
\details{
The index values will be checked for potential errors prior to reporting and flagged if necessary:
\itemize{
 \item DATA_01: Missing data required to calculate the index
 \item DATA_03: Unable to calculate ratio because denominator was zero}
}
\examples{
#calculate the ratio of Peak A to Peak T
pA <- get_fluorescence(example_eems, ex=250:260, em=380:480)
pT <- get_fluorescence(example_eems, ex=270:280, em=320:350)
rAT <- get_ratios(pA, pT)

}
