% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample-info.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Extract components from an \code{eemlist} or \code{abslist}}
\usage{
get_sample_info(x, info)
}
\arguments{
\item{x}{An \code{eemlist}, \code{eem}, \code{abslist}, or \code{abs} object.}

\item{info}{The name of the component to extract.
See \link[eemR:eem]{eemR::eem} for base \code{eem} component names,
\code{\link[=abs_read]{abs_read()}} for base \code{abs} component names,
and \code{\link[=add_metadata]{add_metadata()}} for extended \code{eem} AND \code{abs} metadata fields.}
}
\value{
if \code{x} is an \code{eemlist} or \code{abslist}:

If \code{x} is an \code{eemlist} or \code{abslist}, the returned structure depends on the
type of the requested component (\code{info}):
\itemize{
\item \strong{vector of length 1}: returns a vector with one value per sample
\item \strong{vector of length >1}: returns a data.frame (samples in columns)
\item \strong{matrix}: returns a list of matrices (one per sample)
}

If \code{x} is a single \code{eem} or \code{abs} object, the corresponding component is
returned as-is, preserving its class and structure.
}
\description{
Helper function building on \code{\link[eemR:eem_names]{eemR::eem_names()}}. Extracts a specified
component from an \code{eem} or \code{abs} object.
}
\examples{
# Get sample names
get_sample_info(example_eems, "sample")
get_sample_info(example_abs, "sample")

# Get analysis dates (metadata required)
eemlist <- add_metadata(metadata, example_eems)
get_sample_info(eemlist, "analysis_date")

# Get DOC for the fifth sample
get_sample_info(eemlist[[5]], "doc_mgL")

# Get emission wavelengths
get_sample_info(example_eems, "em")

# Get absorbance matrices
get_sample_info(example_abs, "data")
}
