% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-SNR.R
\name{get_SNR}
\alias{get_SNR}
\title{Title}
\usage{
get_SNR(eem, method = "sqrt")
}
\arguments{
\item{eem}{an \code{eem} or \code{eemlist} object containing EEM's data}

\item{method}{method used to get signal to noise, either a custom function or
"sqrt", the default. See \link[eemanalyzeR]{calc_raman_SNR_sqrt} for details.}
}
\value{
a vector of signal to noise ratios for each sample in the \code{eem} or \code{eemlist}
}
\description{
Title
}
\examples{
eemlist <- add_metadata(metadata, example_eems)
eemlist <- add_blanks(eemlist, validate=FALSE)
SNR <- get_SNR(eemlist)

}
