% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{load_config}
\alias{load_config}
\title{Load user defaults}
\usage{
load_config(
  config_path = rappdirs::user_data_dir("eemanalyzeR"),
  env = .pkgenv
)
}
\arguments{
\item{config_path}{path the YAML file with user default values}

\item{env}{the environment name to write to}
}
\description{
Load user defaults
}
