% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longterm_blank}
\alias{longterm_blank}
\title{Long Term Average Blank}
\format{
an \code{eem} object, for more details on the format of an eem, see the \link[eemR]{eemR} package.
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022 to 2024)
}
\usage{
longterm_blank
}
\description{
An \code{eem} object containing the average of 81 blank samples, used to check for blank consistency.
}
\note{
for details on creating your own long-term blank see ##to do: make function or vignette here
}
\keyword{datasets}
