% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-qaqc.R
\name{mark_qaqc}
\alias{mark_qaqc}
\title{Specify QAQC samples}
\usage{
mark_qaqc(x, blk_pattern = NULL, tea_pattern = NULL, info = "sample")
}
\arguments{
\item{x}{an object of class `eemlist` or `abslist`}

\item{blk_pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the instrument blanks.}

\item{tea_pattern}{a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the tea or check standard.}

\item{info}{the name of the component within the \code{eem} to check for the pattern. default is 'sample'}
}
\value{
an object of the same class as `x`
}
\description{
Updates an `eemlist` or `abslist` with information about which samples are
**instrument** blanks or tea check standards. Marks `eem` or `abs` attributes `is_blank` and `is_check_std` as true.
}
\examples{
eemlist <- mark_qaqc(example_eems, blk_pattern ="BEM", tea_pattern="Tea")
get_sample_info(eemlist[[1]], "sample")
attributes(eemlist[[1]])
get_sample_info(eemlist[[4]], "sample")
attributes(eemlist[[4]])
}
