% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tea-absorbance.R
\name{plot_absorbance_error}
\alias{plot_absorbance_error}
\title{Plot the absorbance of a failed tea absorbance validation check.}
\usage{
plot_absorbance_error(model_data, sample_abs_vector, sample_id, condition)
}
\arguments{
\item{model_data}{tea absorbance model data}

\item{sample_abs_vector}{tea sample absorbance vector}

\item{condition}{was the failure a warning or an error?}

\item{sample_name}{name of the tea sample to use as title of plot}
}
\value{
plot of tea sample absorbance and model of "good" absorbance
}
\description{
`validate_tea_absorbance()` uses this function to plot the tea absorbance
validation model vs. the sample tea absorbance that is outside the
}
