% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eem.R
\name{plot_eem}
\alias{plot_eem}
\title{Plot excitation emission matrices with ggplot2}
\usage{
plot_eem(eem, nbin = 8, equal_scale = FALSE, pal = NULL)
}
\arguments{
\item{eem}{an \code{eem} or \code{eemlist} object containing EEM's data}

\item{nbin}{the number of bins used in the contour plot}

\item{equal_scale}{logical, should the scale be the same for all the plots in the eemlist?}

\item{pal}{the colors used for the fill scale, if not specified it will use the \link[pals]{parula} palette.
If less colors are provided than required, it will use \link[grDevices]{colorRampPalette} to fill in colors.}
}
\value{
a ggplot2 object
}
\description{
Used to make a nice plot of one or multiple excitation emission matrices (EEMs) using ggplot2
}
\details{
The object returned is a \code{ggplot2} object, so it is compatible with other ggplot2 functions to
further modify a plot as desired.
}
\examples{

eems <- add_metadata(metadata, example_eems)
eems <- add_blanks(eems)
eems <- subtract_blank(eems)
eems <- remove_scattering(eems)

#plot just one EEM
plot_eem(eems[[3]])

#plot all EEMs in an eemlist
plot_eem(eems)

#change color scale
plot_eem(eems, pal=c("darkblue", "lightblue"))

#make color bar consistent across all plots
plot_eem(eems, equal_scale=TRUE)

#customize using ggplot2 commands
plot_eem(eems[[2]]) + ggplot2::labs(title="Test EEM")

}
\seealso{
\href{https://ggplot2.tidyverse.org}{ggplot2}
}
