% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readme.R
\name{print_readme}
\alias{print_readme}
\title{Print `readme`}
\usage{
print_readme()
}
\description{
Nicely print the processing documentation readme file with proper formatting.
The readme tracks the processing steps applied to the dataset and is stored
as an object called `readme` in the package environment.
}
