% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-env.R
\name{reset_config}
\alias{reset_config}
\title{Reset all eemanalyzeR settings to package defaults}
\usage{
reset_config(env = .pkgenv)
}
\arguments{
\item{env}{the environment that stores the processing settings. Defaults to the package environment.
It is not recommended the user modifies this argument.}
}
\value{
Invisibly returns the reset default configuration settings as a named list.
}
\description{
Returns the data processing configuration settings for this R session back to package defaults.
}
\details{
This allows the user to return the data processing settings back to the default configuration of the
eemanalyzeR package. These defaults are documented in data.R under "default_config". This function is
provided in case the user needs to return back to default processing settings after experimenting with
modifying the settings using `modify_config`.
}
\examples{
# Reset the configuration back to package defaults
reset_config()
}
