% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-validation-models.R
\name{save_absorbance_model}
\alias{save_absorbance_model}
\title{Create a long term absorbance or EEM model (for tea stds and blanks)}
\usage{
save_absorbance_model(
  abs_dir,
  save_path = "data",
  sd_multiplier = 3,
  overwrite = FALSE
)
}
\arguments{
\item{abs_dir}{location of absorbance data. This should be a directory with multiple
absorbance data files (something like )}

\item{save_path}{location to save absorbance model as a CSV}

\item{sd_multiplier}{acceptable standard deviation in data}

\item{overwrite}{overwrite exisiting model?}
}
\value{
a absorbance model of averaged absorbance
}
\description{
Create a long term absorbance or EEM model (for tea stds and blanks)
}
