% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.abslist}
\alias{[.abslist}
\title{Subsetting using \code{[} for an abslist}
\usage{
\method{[}{abslist}(abslist, i)
}
\arguments{
\item{abslist}{the abslist to subset}

\item{i}{the index for subsetting}
}
\value{
an object of class \code{abslist}
}
\description{
Subsetting using \code{[} for an abslist
}
\keyword{internal}
