% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[[.eemlist}
\alias{[[.eemlist}
\title{Subsetting using \code{[[} for eemlist - we want this to return an eem}
\usage{
\method{[[}{eemlist}(eemlist, i)
}
\arguments{
\item{eemlist}{the eemlist to subset}

\item{i}{the index for subsetting}
}
\value{
an object of class \code{eem}
}
\description{
Subsetting using \code{[[} for eemlist - we want this to return an eem
}
\keyword{internal}
