% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset-eems}
\alias{subset-eems}
\alias{eem_get_info}
\alias{eem_select}
\title{Subset eemlist using components}
\usage{
eem_get_info(eem, info)

eem_select(eem, info, sample, keep = F, ignore_case = F, verbose = T)
}
\arguments{
\item{eem}{an object of class \code{eem} of \code{eemlist}}

\item{info}{the name of the component within the \code{eem} to extract.
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{add_meta}
for extended \code{eem} component names.}

\item{sample}{a vector of the names or other info to use to select EEM's from \code{eemlist}}

\item{keep}{logical. If TRUE, the specified sample will be returned. If FALSE, they will be removed.}

\item{ignore_case}{logical, should sample name case should be ignored (TRUE) or not (FALSE). Default is FALSE.}

\item{verbose}{logical determining if removed/extracted eems should be printed on screen.}
}
\value{
\code{eem_get_info} returns a vector containing the info from the EEM's.

\code{eem_select} returns an object of class \code{eemlist} with the samples excluded or selected based on the 'sample' vector.
}
\description{
These are helper functions that build upon the \link[eemR]{eem_names} and \link[eemR]{eem_extract}
functions. \code{eem_get_info} will extract components of an \code{eem} object. While
\code{eem_select} will select or remove samples based on the info extracted by \code{eem_get_info}.
These functions allow selection based on components besides just the sample name.
}
\examples{
#get names
eem_get_info(example_eems, "sample")

#get analysis_date
eemlist <- eem_add_meta(metadata, example_eems)
eem_get_info(eemlist, "analysis_date")

#get doc
eemlist <- eem_add_meta(metadata, example_eems)
eem_get_info(eemlist, "doc_mgL")

#subset by name
names <- eem_get_info(example_eems, "sample")
eem_subset <- eem_select(example_eems, "sample", names[1]) #default is to remove
eem_subset <- eem_select(example_eems,"sample", names[1], keep=T) #but use keep to keep instead

#subset by file_name
eemlist <- eem_add_meta(metadata, example_eems)
names <- eem_get_info(eemlist, "file_name")
eem_subset <- eem_select(eemlist, "file_name", names[1]) #default is to remove

}
