% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-qaqc.R
\name{subset_qaqc}
\alias{subset_qaqc}
\title{Extract QAQC samples from an eemlist or abslist}
\usage{
subset_qaqc(x, type = c("blank", "tea_std"), negate = FALSE)
}
\arguments{
\item{x}{an object of class \code{eemlist} or \code{abslist}}

\item{type}{either "blank" to return instrument blanks or "tea_std" to return tea check standard}

\item{negate}{logical, should the QAQC samples be returned or removed?}
}
\value{
an object of the same class as \code{x} with only the blanks or tea standards.
}
\description{
Extracts either \strong{instrument} blanks or tea check standards from an \code{eemlist} or \code{abslist}
using the samples attributes. Samples labeled using the the \link[eemanalyzeR]{mark_qaqc} function.
}
\examples{
#returns NULL because no samples are marked as QAQC
tea <- subset_qaqc(example_abs, "tea_std")
tea

#get tea absorbance
abs <- mark_qaqc(example_abs, tea_pattern="Tea")
tea <- subset_qaqc(abs, "tea_std")
get_sample_info(tea, "sample")

#get blank eems
eems <- mark_qaqc(example_eems, blk_pattern="BEM", tea_pattern="Tea")
blk <- subset_qaqc(eems)
get_sample_info(blk, "sample")

#use negate = TRUE to get non qaqc samples
nonblk <- subset_qaqc(eems, negate=TRUE)
get_sample_info(nonblk, "sample")
}
