% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-samples.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset \code{eemlist} or \code{abslist} based on components}
\usage{
subset_samples(
  x,
  info,
  sample,
  keep = FALSE,
  ignore_case = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{An \code{eemlist} or \code{abslist} object.}

\item{info}{Name of the component within the \code{eem} or \code{abs} to extract. See
\link[eemR:eem]{eemR::eem} for base \code{eem} components and \code{\link[=abs_read]{abs_read()}}
for base \code{abs} components. Extended components may be added with \code{\link[=add_metadata]{add_metadata()}}.}

\item{sample}{A vector of names or other component values to select/exclude from \code{x}.}

\item{keep}{Logical. If \code{TRUE}, returns the specified samples; if \code{FALSE}, removes them. Default is \code{FALSE}.}

\item{ignore_case}{Logical. Should case be ignored when matching? Default is \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE}, prints removed or extracted samples to the console.}
}
\value{
An object of class \code{eemlist} or \code{abslist} with samples selected or removed based on \code{sample}.
}
\description{
Helper function to select or remove samples from an \code{eemlist} or \code{abslist} using components
beyond just the sample name. Builds on \code{\link[eemR:eem_extract]{eemR::eem_extract()}}.
}
\examples{
# Subset by sample name
names <- get_sample_info(example_eems, "sample")
eem_subset <- subset_samples(example_eems, "sample", names[1]) # removes by default
eem_subset <- subset_samples(example_eems, "sample", names[1], keep = TRUE) # keeps instead

# Subset by metadata name
eemlist <- add_metadata(metadata, example_eems)
names <- get_sample_info(eemlist, "sample_id")
eem_subset <- subset_samples(eemlist, "sample_id", names[1]) # removes by default
}
\seealso{
\code{\link[=get_sample_info]{get_sample_info()}}
}
