% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-type.R
\name{subset_type}
\alias{subset_type}
\title{Extract Samples from an eemlist or abslist by sample type}
\usage{
subset_type(x, type = c("iblank", "sblank", "check", "sample"), negate = FALSE)
}
\arguments{
\item{x}{an object of class \code{eemlist} or \code{abslist}}

\item{type}{either 'iblank' for instrument blank, 'sblank' for analytical blank, 'check' for tea check standard, or 'sample' for samples}

\item{negate}{logical, should the samples be returned or removed?}
}
\value{
an object of the same class as \code{x}.
}
\description{
Extracts the specified sample type(s) from an \code{eemlist} or \code{abslist}
using the samples attribute \code{sample_type}. Samples labeled using the the \link[eemanalyzeR]{add_metadata} function.
}
\examples{
abs <- add_metadata(metadata, example_abs)
eem <- add_metadata(metadata, example_eems)

#returns NULL because no samples are marked as instrument blanks
tea <- subset_type(abs, "iblank")
tea

#get tea absorbance
tea <- subset_type(abs, "sblank")
get_sample_info(tea, "sample")

#get blank eems (instrument and analytical)
blk <- subset_type(eem, c("iblank", "sblank"))
get_sample_info(blk, "sample")

#use negate = TRUE to get non instrument blanks
nonblk <- subset_type(eem, "iblank", negate=TRUE)
get_sample_info(nonblk, "sample")
}
