% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract-blank.R
\name{subtract_blank}
\alias{subtract_blank}
\title{Perform blank subtraction on EEM samples}
\usage{
subtract_blank(eem)
}
\arguments{
\item{eem}{An \code{eem} or \code{eemlist} object.}
}
\value{
An object of class \code{eem} or \code{eemlist} where the sample data (\code{x})
has been subtracted by the blank data (\code{blk_x}).
}
\description{
Subtracts the blank EEM data from sample EEM data. Samples must have
metadata and blank data attached prior to running.
}
\details{
Blank subtraction requires that blank EEM data be attached to each sample.
This is done with \code{\link[=add_blanks]{add_blanks()}}, which in turn requires metadata to be added
via \code{\link[=add_metadata]{add_metadata()}}.
}
\examples{
eem <- add_metadata(metadata, example_eems)
blanklist <- subset_type(eem, "iblank")
eem <- add_blanks(eem, blanklist)

# Subtract blank from a single EEM
eem_sub <- subtract_blank(eem[[1]])

# Subtract blank from an EEM list
eemlist_sub <- subtract_blank(eem)
}
\seealso{
\code{\link[=add_metadata]{add_metadata()}}
}
