% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tea_absorbance_model}
\alias{tea_absorbance_model}
\title{Tea Absorbance Model}
\format{
tibble with columns: wavelength, mean_abs_by_wavelength, sd_abs_by_wavelength, sdmin_3x, sdmax_3x
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022 to 2024)
}
\usage{
tea_absorbance_model
}
\description{
An tibble containing the average absorbance of multiple 1% Pure Leaf Unsweetened Black Tea (SRMtea) standards.
}
\note{
for details on creating your own long-term SRMtea see ##to do: make function or vignette here
}
\keyword{datasets}
