% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique-samples}
\alias{unique-samples}
\alias{unique.eemlist}
\alias{unique.abslist}
\title{Get unique EEM's or absorbance}
\usage{
\method{unique}{eemlist}(x, ...)

\method{unique}{abslist}(x, ...)
}
\arguments{
\item{x}{an object of class \code{eemlist} or \code{abslist}}

\item{...}{additional arguments passed to \link[base]{unique}}
}
\value{
an object of class \code{eemlist} or \code{abslist} with samples with duplicate EEM's matrices (eem$x) or absorbance data (abs$data) removed.

Note: sample names remain unchanged.
}
\description{
Get unique EEM's or absorbance
}
\examples{
unique_eems <- unique(example_eems)
unique_abs <- unique(example_absorbance)
}
