% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample.R
\name{upsample}
\alias{upsample}
\alias{upsample_eem}
\alias{upsample_abs}
\title{Upsample EEMs and Absorbance data}
\usage{
upsample_eem(eemlist, interval = 1, method = 1, cores = 1)

upsample_abs(abslist, interval = 1)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data}

\item{interval}{numeric, the scale at which the EEM should be upscaled to. For example: 1 will provide an interpolated EEM with values at every 1 nm.}

\item{method}{numeric 0 to 4 for interpolation method to use. Default is 1. See \link[staRdom]{eem_interp} for more info.}

\item{cores}{The number of cores used for parallel computation of interpolation.}

\item{abslist}{an \code{abslist} object containing absorbance data}
}
\value{
an \code{eemlist} or \code{abslist} with wavelengths interpolated
}
\description{
It is helpful when finding indices to have a more detailed data with wavelengths every 1 nanometer, this function can be used to
interpolate an eemlist or abslist to 1 nm intervals (or other interval) for this purpose.
}
\note{
Interpolation of the absorbance data uses spline interpolation via the \link[zoo]{na.spline} function.
}
\examples{
eemlist <- example_eems
eemlist <- remove_scattering(eemlist)
eemlist_upscale <- upsample_eem(eemlist)
plot_eem(eemlist_upscale)

abslist <- example_absorbance
abslist_upscale <- upsample_abs(abslist)
}
