% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs-indices.R
\name{usgs_indices}
\alias{usgs_indices}
\title{USGS methods for fluorescence and absorbance indices}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
\emph{Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance
and fluorescence from dissolved organic matter} (USGS Numbered Series No. 2018-1096).
U.S. Geological Survey. \doi{10.3133/ofr20181096}
}
\usage{
usgs_indices(eemlist, abslist, cuvle = 1, qaqc_dir = NA)
}
\arguments{
\item{eemlist}{An \code{eemlist} object.}

\item{abslist}{An \code{abslist} object.}

\item{cuvle}{Cuvette (path) length in cm.}

\item{qaqc_dir}{File path to the QAQC files generated with \code{\link[=create_mdl]{create_mdl()}} and \code{\link[=create_std]{create_std()}}.
Default is NA, which results in NO QAQC CHECKS.}
}
\value{
A list with two elements:
\itemize{
\item \strong{eem_index}: a \code{data.frame} of all fluorescence indices. Each row corresponds to a single index for a sample.
\item \strong{abs_index}: a \code{data.frame} of all absorbance indices. Each row corresponds to a single index for a sample.
}

Each \code{data.frame} contains the following columns:
\itemize{
\item \strong{sample_id}: sample identifier for the sample (if provided, otherwise uses sample: the sample's file name)
\item \strong{sample_name}: sample name or description (if provided, otherwise uses sample: the sample's file name)
\item \strong{index}: name of the index
\item \strong{value}: calculated value of the index
}
}
\description{
Calculates commonly used absorbance and fluorescence optical indices from \code{eemlist} and \code{abslist}.
These indices are those commonly used by the U.S. Geological Survey.
For detailed descriptions and references, see Hansen et al. 2018 (Tables 1 and 8).
}
\note{
\itemize{
\item If absorbance is not at a 1 nm interval, it will be interpolated using \code{\link[zoo:na.approx]{zoo::na.approx()}}, which fills in missing values using linear interpolation.
\item If EEM data is not at a 1 nm interval, fluorescence will be interpolated using \code{\link[pracma:interp2]{pracma::interp2()}}.
}
}
\examples{
indices <- usgs_indices(
  example_processed_eems,
  example_processed_abs,
  qaqc_dir = system.file("extdata", package = "eemanalyzeR")
)
}
