% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-blanks.R
\name{validate_blanks}
\alias{validate_blanks}
\title{Validate the instrument blank(s)}
\usage{
validate_blanks(blanklist)
}
\arguments{
\item{blanklist}{An \code{eemlist} containing containing the blank EEMs.}
}
\value{
A blanklist with validated blanks, or an error if no valid blanks are selected
}
\description{
Plots validation instrument blank samples for manual inspection.
}
\examples{
eems <- add_metadata(metadata, example_eems)
eems <- subset_type(eems, type = "iblank")
valid_blanklist <- validate_blanks(eems)
}
