% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-env.R
\name{validate_config}
\alias{validate_config}
\title{Validate the eemanalyzeR configuration}
\usage{
validate_config(env = .pkgenv)
}
\arguments{
\item{env}{environment where to find the config. Defaults to the package environment}
}
\description{
Checks that all settings in the eemanalyzeR config are valid options and warns user if not
}
\examples{
# Example validation
validate_config()
}
