% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-std.R
\name{validate_std}
\alias{validate_std}
\title{Visually validate the check standard}
\usage{
validate_std(abslist, qaqc_dir = get_qaqc_dir(), tolerance = 0.2)
}
\arguments{
\item{abslist}{An \code{abslist} object.}

\item{qaqc_dir}{File path to the QAQC files generated with \code{\link[=create_mdl]{create_mdl()}} and \code{\link[=create_std]{create_std()}}.
Default is a user-specific data directory \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}.}

\item{tolerance}{Maximum percent deviation that the check standard can vary
from the long-term values without being flagged.}
}
\value{
A \code{ggplot} object showing the absorbance of the check standards,
the long-term standard (dashed line), and the tolerance thresholds (gray ribbon).
}
\description{
Visually inspect plots of the tea check standard absorbance compared to the long-term standard.
}
\examples{
abslist <- add_metadata(metadata, example_abs)
validate_std(abslist, system.file("extdata", package = "eemanalyzeR"))
}
\seealso{
\code{\link[=create_std]{create_std()}}
}
