% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tea-absorbance.R
\name{validate_tea_absorbance}
\alias{validate_tea_absorbance}
\title{Validate Tea Absorbance}
\usage{
validate_tea_absorbance(
  abslist,
  pattern = "tea",
  model = "default",
  fail_threshold = 0.1,
  ...
)
}
\arguments{
\item{abslist}{abslist of sample absorbances}

\item{pattern}{string which is used to extract tea samples from all absorbance.
User can modify this based on each lab's sample ID scheme. Defaults to "tea".}

\item{model}{the comparison model of tea absorbance}

\item{fail_threshold}{user adjustable threshold to determine failure of the
tea absorbance validation procedure. Must be a decimal between 0 and 1. The value
is the minimum percent of the absorbance curve that is outside the error bar
for validation failure. For example if the value is 0.15, then the check fails
if > 15% of the sample absorbance curve is outside the modeled error bars.
Defaults to 0.10.}

\item{...}{further arguments to pass to ggplot for to control plotting the tea
absorbance model}
}
\value{
returns a data.frame with pass/fail results of tea standard checks
}
\description{
Validates the absorbance of sample tea standards compared to a mean of good
tea standards run in the past on the instrument. Returns an error if the sample
absorbance curve is outside the fail threshold, and a warning if inside the fail
threshold but not fully within historical error bars.
}
\section{Default tea absorbance model}{


If model = "default", the model included in the package is used to validate
absorbance. The included model is the repeated absorbance of a 1% Unsweetened
Pure Leaf Black Tea solution created following USGS SRMtea method. See
https://pubs.usgs.gov/of/2018/1096/ofr2018.1096.pdffor further explanation of
SRMtea standard.

...
}

