% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{validate_user_config}
\alias{validate_user_config}
\title{Validate the eemanalyzeR configuration}
\usage{
validate_user_config(user_config_file = .user_config_path())
}
\arguments{
\item{user_config_file}{The path to the file that stores the user config. Defaults to the yaml file in the default user data directory.}
}
\value{
invisibly returns TRUE if the configuration is valid, otherwise returns an error
}
\description{
Checks that all settings in the eemanalyzeR user config are valid options that match the 
template included with the package.
}
\examples{
# Example validation
validate_user_config()
}
