% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{abs_dir_read}
\alias{abs_dir_read}
\title{Read absorbance data files from directory}
\usage{
abs_dir_read(
  input_dir,
  pattern = NULL,
  skip = "SEM|BEM|waterfall",
  file_ext = "dat",
  recursive = FALSE
)
}
\arguments{
\item{input_dir}{path to folder containing raw absorbance files}

\item{pattern}{optional. a case-sensitive character string to be matched to the files in input_dir.
only files matching the pattern will be loaded. to use multiple strings, use the form "str1|str2".}

\item{skip}{a case-insensitive character string to be matched to the files in input_dir.
any files matching this string will be ignored. useful for ignoring absorbance data.
to use multiple strings, use the form "str1|str2".}

\item{file_ext}{character. the file extension of the EEMs}

\item{recursive}{logical. should the function recurse into directories?}
}
\value{
a data.frame where the first column is the wavelengths the absorbance was measured at and each
subsequent column is the absorbance for a sample where the column name is the sample name.
}
\description{
A flexible function that will read all absorbance files in a directory even when
it contains other files.
}
\examples{
#load samples by using pattern to identify absorbance samples
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "abs|ABS")

#load samples by using skip to exclude EEM's and other samples
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), skip = "SEM|BEM|waterfall")


}
