% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-interp.R
\name{abs_interp}
\alias{abs_interp}
\title{Interpolate absorbance data}
\usage{
abs_interp(abs, type = "linear")
}
\arguments{
\item{abs}{an object of class \code{abs} or \code{abslist}}

\item{type}{the method used to interpolate, either "linear" for linear interpolation using \link[zoo]{na.approx} or "spline" for spline interpolation using \link[zoo]{na.spline}}
}
\value{
an object of class \code{abs} or \code{abslist}
}
\description{
Interpolates absorbance data to 1 nanometer resolution to enable index calculation at
wavelengths that weren't specifically measured.
}
\examples{
abslist_filled <- abs_interp(example_abs)
abs_filled <- abs_interp(example_abs[[1]])
}
