% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{abs_read}
\alias{abs_read}
\title{Read a single absorbance file into R}
\usage{
abs_read(file)
}
\arguments{
\item{file}{file path to the absorbance file to load}
}
\value{
a data.frame with 2 columns where the first column is the absorbance wavelength and the second column
is the absorbance measured.
}
\description{
Will accept absorbance files run via Aqualog exported using the sample Q or manually. If the function tries
to load an EEM's file it will return a warning and a NULL value for absorbance.
}
\examples{
abs_files <- list.files(system.file("extdata", package = "eemanalyzeR"),
full.names=TRUE, pattern="ABS")
abs <- abs_read(abs_files[1])
}
