% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{abs_read}
\alias{abs_read}
\title{Read a single absorbance file into R}
\usage{
abs_read(file)
}
\arguments{
\item{file}{file path to the absorbance file to load}
}
\value{
An object of class \code{abs} containing:
\itemize{
 \item file The filename of the absorbance data.
 \item sample The sample name of the absorbance data.
 \item n The number of wavelengths absorbance was measured at.
 \item data A \code{data.frame} with absorbance data.
 \item dilution The dilution factor for the absorbance data.
 \item analysis_date The date the sample was run.
 \item description Optional description of sample.
 \item doc_mgL The concentration of dissolved organic carbon in the sample given in mg \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}
 \item notes Optional notes related to the sample or sample collection.
 \item location Directory of the absorbance data.
}
}
\description{
Will accept absorbance files run via Aqualog exported using the sample Q or manually. If the function tries
to load an EEM's file it will return a warning and a NULL value for absorbance.
}
\note{
\code{abs} attributes dilution through notes will remain empty until merged with metadata in \link[eemanalyzeR]{abs_add_meta}
}
\examples{
abs_files <- list.files(system.file("extdata", package = "eemanalyzeR"),
full.names=TRUE, pattern="ABS")
abs <- abs_read(abs_files[1])
}
