% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-read.R
\name{abs_read}
\alias{abs_read}
\title{Read a single absorbance file into R}
\usage{
abs_read(file)
}
\arguments{
\item{file}{file path to the absorbance file to load}
}
\value{
An object of class \code{abs} containing:
\itemize{
 \item file The filename of the absorbance data.
 \item sample The sample name of the absorbance data.
 \item n The number of wavelengths absorbance was measured at.
 \item data A matrix where the first column is the wavelengths in nm and the second column is the absorbance.
 \item location Directory of the absorbance data.
}
}
\description{
Will accept absorbance files run via Aqualog exported using the sample Q or manually. If the function tries
to load an EEM's file it will return a warning and a NULL value for absorbance.
}
\examples{
abs_files <- list.files(system.file("extdata", package = "eemanalyzeR"),
full.names=TRUE, pattern="ABS|Abs ?Spectra ?Graphs")
abs <- abs_read(abs_files[1])
}
