% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-blanks.R
\name{add_blanks}
\alias{add_blanks}
\title{Check blanks and add to sample data}
\usage{
add_blanks(eemlist, blanklist = NULL, pattern = "BEM|Blank$", validate = TRUE)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{blanklist}{optional. an \code{eem} or \code{eemlist} containing the EEM's data for the blank(s)}

\item{pattern}{optional. a character string containing a \code{\link[base]{regular expression}}
used to specify the names of the samples to be used for blanks.}

\item{validate}{logical. If TRUE, will print out the blanks and ask user for validation.
If FALSE, it will skip validation and just add the blanks}
}
\value{
an \code{eemlist} where each \code{eem} object has two additional items:
\itemize{
\item \code{blk_file}: the file location for the blank associated with the sample
\item \code{blk_x}: the blank EEM associated with the sample
}
}
\description{
Allows user to visually check excitation emissions matrix plots of the blanks, and then, if approved, the blanks will
be added to the \code{eem} objects for use in further processing.
}
\details{
If more than one blank is supplied, the function links the blank and sample by the metadata
name which should be the same between the sample and the blank, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.

Adding the blanks into the sample data can be done two ways:
\enumerate{
 \item Include an \code{eemlist} containing EEM's data for both sample and blank: Need to specify a \code{pattern} to identify blanks
 \item Include an \code{eemlist} containing \strong{only} EEM's data for samples: Need to include an \code{eem} or \code{eemlist} as \code{blanklist} with EEM's data for blank(s)
}
}
\examples{
eemlist <- add_metadata(metadata, example_eems)
augment_eemlist <- add_blanks(eemlist, validate=FALSE)

}
