% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to absorbance and EEM's data}
\usage{
add_metadata(meta, x, blank_pattern = "BEM|BLK|blank", check_pattern = "tea")
}
\arguments{
\item{meta}{a \code{data.frame} of metadata}

\item{x}{an \code{abslist} or \code{eemlist} object}

\item{blank_pattern}{a regex string to find blanks using pattern matching on filenames}

\item{check_pattern}{a regex string to find check standards using pattern matching on filenames}
}
\value{
returns an object of the same class as the object input into the function (\code{abslist} or \code{eemlist}) with metadata added in
The following items are added for each sample, if available in the metadata:
\itemize{
 \item meta_name: the data identifier of the sample.
 \item dilution: the dilution factor for the sample.
 \item analysis_date: the date the sample was run.
 \item description: optional description of sample.
 \item is_blank: logical whether the sample is a blank (e.g. MilliQ water)
 \item is_check: logical whether the sample is a check std (e.g SRM Tea)
 \item doc_mgL: the concentration of dissolved organic carbon in the sample given in mg \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}
 \item notes: optional notes related to the sample or sample collection.
}
if \code{x} is a \code{eemlist} an additional two items are added to each sample:
\itemize{
 \item integration_time_s: the integration time for the sample.
 \item raman_area_1s: the area under the raman water peak with a 1 second integration time.
}
}
\description{
Add metadata to absorbance and EEM's data
}
\note{
if \code{eemlist} contains blanks, the blanks will get the metadata of the corresponding sample.
}
\examples{

#add metadata to absorbance data
abs_augment <- add_metadata(metadata, example_absorbance)

#add metadata to EEM's data
eem_augment <- add_metadata(metadata, example_eems)
}
