% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{annotate_plot}
\alias{annotate_plot}
\title{Annotate EEM plot with location of indices}
\usage{
annotate_plot(plot, index_method = "eemanalyzeR")
}
\arguments{
\item{plot}{A \code{ggplot2} object of an \code{eem}.}

\item{index_method}{Either "eemanalyzeR", "eemR", "usgs".}
}
\value{
A \code{ggplot2} object.
}
\description{
Adds labels indicating the regions of an excitation–emission matrix (EEM)
that are used to calculate fluorescence indices. The annotation is
customized according to the selected index method.
}
\examples{
plot <- plot(example_processed_eems[[3]])
annotate_plot(plot)
}
