% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mdl.R
\name{check_eem_mdl}
\alias{check_eem_mdl}
\alias{check_abs_mdl}
\title{Check if a EEM sample is above the MDL}
\usage{
check_eem_mdl(eem, mdl = NULL, ex, em, vals = FALSE)

check_abs_mdl(abs, mdl = NULL, wl, vals = FALSE)
}
\arguments{
\item{eem}{an object of class \code{eem} or \code{eemlist}}

\item{mdl}{a \code{eem} object containing MDL data, if no MDL provided will return NA}

\item{ex}{a vector of excitation wavelengths}

\item{em}{a vector of emission wavelengths}

\item{vals}{logical, if FALSE will return a TRUE or FALSE, if TRUE will
return a table with the eem and mdl values for the ex/em pair}

\item{abs}{an object of class \code{abs} or \code{abslist}}

\item{wl}{a vector of absorbance wavelengths}
}
\value{
If \code{vals} is FALSE:
\itemize{
\item MDL01 if all values are below the MDL
\item MDL02 if some of the values are below the MDL
\item NA if all are above the MDL
}

If \code{vals} is TRUE:
\itemize{
\item a \code{data.frame} with four columns:
\itemize{
\item ex: the excitation wavelengths
\item em: the emission wavelengths
\item eem: the EEM fluorescence value for the ex/em pair
\item mdl: the MDL for the ex/em pair
}
}
}
\description{
Given a set of excitation and emission wavelengths, will check if the values in the
\code{eem} or \code{eemlist} are above the method detection limit (MDL)
for all excitation-emission pairs.
}
\examples{
#get MDL data
eem_mdl <- readRDS(file.path(system.file("extdata", package = "eemanalyzeR"),
"eem-mdl.rds"))
abs_mdl <- readRDS(file.path(system.file("extdata", package = "eemanalyzeR"),
"abs-mdl.rds"))

#works with a single sample
check_eem_mdl(example_processed_eems[[1]], eem_mdl, ex = 270:280, em=300:320)

#or an eemlist
check_eem_mdl(example_processed_eems, eem_mdl, ex = 270:280, em=300:320)

check_abs_mdl(example_processed_abs[[2]], abs_mdl, wl=254)
check_abs_mdl(example_processed_abs, abs_mdl, wl=254)
}
