% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-tea.R
\name{check_tea_std}
\alias{check_tea_std}
\title{Check if a tea standard is consistent with long-term standard}
\usage{
check_tea_std(
  eemlist,
  abslist,
  std_dir = .qaqc_dir(),
  tolerance = 0.2,
  index_method = "eemanalyzeR",
  vals = FALSE
)
}
\arguments{
\item{eemlist}{an object of class \code{eemlist}, needs to be fully processed}

\item{abslist}{an object of class \code{abslist}, needs to be fully processed}

\item{std_dir}{file path to the QAQC files generated with \link[eemanalyzeR]{create_mdl} and \link[eemanalyzeR]{create_tea_std},
default is a user-specific data directory (\link[rappdirs]{user_data_dir})}

\item{tolerance}{what is the maximum percentage the tea standard can vary from the long-term values without being flagged?}

\item{index_method}{the index method used to calculate indices, see \link[eemanalyzeR]{get_indices} for more info.}

\item{vals}{logical, if FALSE will return a flag, if TRUE will
return a table with the observed and mdl values for the ex/em pair}
}
\value{
#TODO: update returns values
If \code{vals} is FALSE:
\itemize{
\item STD01 if all values outside of the tolerance
\item STD02 if some of the values are outside of the tolerance
\item NA if all are within the tolerance
}

If \code{vals} is TRUE:
\itemize{
\item a \code{data.frame} with the index, the observed value, the long-term standard value, and the percent deviation
}
}
\description{
Calculate the indices in a tea standard to ensure they are consistent
with the long-term standard.
}
\examples{
check_tea_std(example_processed_eems, example_processed_abs,
              std_dir = system.file("extdata", package = "eemanalyzeR"))
}
