% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-tea-std.R
\name{check_tea_std}
\alias{check_tea_std}
\title{Check if a tea standard is consistent with long-term standard}
\usage{
check_tea_std(
  eemlist,
  abslist,
  std_dir = .qaqc_dir(),
  tolerance = 0.2,
  index_method = "eemanalyzeR",
  vals = FALSE
)
}
\arguments{
\item{eemlist}{an object of class \code{eemlist}, needs to be fully processed}

\item{abslist}{an object of class \code{abslist}, needs to be fully processed}

\item{std_dir}{file path to the QAQC files generated with \link[eemanalyzeR]{create_mdl} and \link[eemanalyzeR]{create_tea_std},
default is a user-specific data directory (\link[rappdirs]{user_data_dir})}

\item{tolerance}{what is the maximum percentage the tea standard can vary from the long-term values without being flagged?}

\item{index_method}{the index method used to calculate indices, see \link[eemanalyzeR]{get_indices} for more info.}

\item{vals}{logical, if FALSE will return a flag, if TRUE will
return a table with the observed and standard values for each index.}
}
\value{
A \code{data.frame} with four columns:
\itemize{
\item meta_name: the metdata name of the tea sample
\item type: either abs or eem to specify the index type
\item index: the name of the index
\item tea_flag: a flag indicating if the index outside the tolerance (STD01 or NA).
}

If \code{vals} is TRUE, the table will also include the observed value, the long-term standard value, and the percent deviation
}
\description{
Calculate the indices in a tea standard to ensure they are consistent
with the long-term standard.
}
\examples{
check_tea_std(example_processed_eems, example_processed_abs,
              std_dir = system.file("extdata", package = "eemanalyzeR"))
}
