% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-dilution.R
\name{correct_dilution}
\alias{correct_dilution}
\title{Perform Dilution Corrections}
\usage{
correct_dilution(x)
}
\arguments{
\item{x}{an \code{eemlist} or \code{abslist} object. See details for more info.}
}
\value{
an object of class \code{eemlist} or \code{abslist}
}
\description{
Adjusts the EEM's and absorbance data fluorescence to account for dilutions performed prior to measurement.
}
\details{
The function uses the dilution factor provided by the metadata to dilution correct the EEMs and absorbance, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.
}
\examples{
eemlist <- add_metadata(metadata,example_eems)
eemlist <- eem_rm_blank(eemlist)
correct_eem <- correct_dilution(eemlist)

abslist <- add_metadata(metadata,example_abs)
abslist <- eem_rm_blank(abslist)
correct_abs <- correct_dilution(abslist)
}
