% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-mdl.R
\name{create_mdl}
\alias{create_mdl}
\title{Calculate method detection limits}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence from dissolved organic matter
(USGS Numbered Series No. 2018-1096). Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence
from dissolved organic matter (Vol. 2018-1096). Reston, VA: U.S. Geological Survey. \url{https://doi.org/10.3133/ofr20181096}
}
\usage{
create_mdl(
  dir,
  meta_name = NULL,
  sheet = NULL,
  iblank = "BEM",
  type = "eem",
  recursive = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{dir}{path to folder containing long term EEMs and/or absorbance files, all files in this directory will be loaded}

\item{meta_name}{name of metadata file. optional if metadata is only xlsx or csv file in input_dir
if not specified function will attempt to load any xlsx or csv file in directory and return an error if there is more than one}

\item{sheet}{name of sheet containing metadata. only required if metadata isn't the first sheet}

\item{iblank}{optional. a character string containing a \code{\link[base]{regular expression}}
used to specify the sample names of the instrument blanks.}

\item{type}{which MDL to calculate: either \code{eem} or \code{abs}}

\item{recursive}{logical. should the function recurse into directories?}

\item{output_dir}{the location to save the mdl file to, default is a user-specific data directory (\link[rappdirs]{user_data_dir}). If
FALSE will return the MDL instead of saving.}
}
\value{
\itemize{
\item If output_dir is FALSE, returns an \code{eem} or \code{abs} object containing the MDL values.
\item Otherwise it saves a .RDS file containing the MDL data formatted as either an \code{eem} or \code{abs} object to the specified output_dir.
}
}
\description{
The method detection limit (MDL) is the signal limit that we can reliably have confidence
that the result is greater than zero and distinguishable from blanks. In this case the MDL
is calculated based on the method proposed by Hansen et al. (2018); three times the standard
deviation plus the mean of the long-term blank.
}
\details{
To calculate the MDL you need:
\itemize{
\item A directory containing analytical blanks with their associated instrument blanks (less than 20 will prompt a warning)
\itemize{
\item Note: sample names must be unique
}
\item Metadata for the blanks including (at a minimum) the integration time and raman area in a single file,
formatted as a metadata file (see \link[eemanalyzeR]{metadata})
}
}
\examples{
eem_mdl <- create_mdl(file.path(system.file("extdata", package = "eemanalyzeR"),
"long-term-blanks"), type="eem", output_dir = FALSE)

plot_eem(eem_mdl)

}
