% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-std.R
\name{create_std}
\alias{create_std}
\title{Create long term average check standard}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
\emph{Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance
and fluorescence from dissolved organic matter} (USGS Numbered Series No. 2018-1096).
U.S. Geological Survey. \doi{10.3133/ofr20181096}
}
\usage{
create_std(
  dir,
  method = "default",
  meta_name = NULL,
  sheet = NULL,
  abs_pattern = "Abs",
  iblank = "BEM",
  type = "eem",
  recursive = FALSE,
  qaqc_dir = get_qaqc_dir(),
  update_config = TRUE
)
}
\arguments{
\item{dir}{Path to a folder containing long-term EEMs and/or absorbance files.
All files in this directory will be loaded.}

\item{method}{A character string describing the method associated with the MDL files.}

\item{meta_name}{Name of the metadata file. Optional if the metadata file is the
only \code{.xlsx} or \code{.csv} file in \code{dir}. If not specified, the function attempts to find
a single metadata file and errors if multiple files are present.}

\item{sheet}{Name of the sheet containing metadata (only required
if the metadata is not on the first sheet of an \code{.xlsx} file).}

\item{abs_pattern}{A character string containing a
\link[base:regex]{base::regular expression} used to identify absorbance files.}

\item{iblank}{Optional. A character string containing a
\link[base:regex]{base::regular expression} used to identify instrument blanks.}

\item{type}{Which MDL to calculate: either "eem" or "abs".}

\item{recursive}{Logical. Should the function recursively search directories?}

\item{qaqc_dir}{Directory in which to save the QAQC \code{.rds} file.
Default: The QAQC directory in the current package environment, usually the user configured QAQC directory.
If \code{NA}, the function returns the MDL object instead of saving it.}

\item{update_config}{Logical. Should function ask to update user_config file with
the default QAQC directory location?}
}
\value{
\itemize{
\item If \code{dir = FALSE}: an \code{eem} or \code{abs} object containing the averaged check standard values.
\item Otherwise: saves an \code{.rds} file containing the averaged check standard and
invisibly returns the file path.
}
}
\description{
Calculates the average absorbance and fluorescence values given multiple check
standards to create a long-term standard that the daily check standard can be
checked against for consistency.
}
\details{
To ensure optical measurements are consistent across days, Hansen et al. (2018)
recommend using a standard reference material: Pure Leaf® unsweetened black tea.
The tea standard should be diluted to 1\% concentration before analysis. It is
recommended to run the tea standard at the start and end of each analysis run.
According to USGS standards, tea standard indices measured on a given day should
fall within 20\% of the long-term average (Hansen et al. 2018).

To calculate the average check standard you need:
\itemize{
\item A directory containing check standards with their associated instrument blanks (less than 20 will prompt a warning)
\itemize{
\item Note: sample names must be unique
}
\item Metadata for the check samples including (at a minimum) the integration time and raman area in a single file,
formatted as a metadata file (see \link{metadata})
}
}
\examples{
eem_std <- create_std(file.path(system.file("extdata", package = "eemanalyzeR"),"long-term-std"),
meta_name="longterm-checkstd-metadata.csv", abs_pattern = "ABS",
type="eem", qaqc_dir = NA, update_config=FALSE)

plot(eem_std)


}
