% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{dir_read}
\alias{dir_read}
\alias{eem_dir_read}
\alias{abs_dir_read}
\title{Read excitation-emission fluorescence matrices (EEM) and absorbance data from directory}
\usage{
eem_dir_read(
  input_dir,
  pattern = NULL,
  skip = "(?i)abs",
  file_ext = "dat",
  recursive = FALSE,
  import_function = "aqualog"
)

abs_dir_read(
  input_dir,
  pattern = NULL,
  skip = "SEM|BEM|Waterfall",
  file_ext = "dat",
  recursive = FALSE
)
}
\arguments{
\item{input_dir}{path to folder containing raw files (either EEM's or absorbance)}

\item{pattern}{optional. a character string containing a \code{\link[base]{regular expression}}
to be matched to the files in input_dir, only files matching the pattern will be loaded.}

\item{skip}{a character string containing a \code{\link[base]{regular expression}} to be matched to the files in input_dir.
any files matching this string will be ignored. useful for ignoring absorbance data.}

\item{file_ext}{character. the file extension of the data files.}

\item{recursive}{logical. should the function recurse into directories?}

\item{import_function}{character or a user-defined function to import an EEM. for more details see \link[eemR]{eem_read}}
}
\value{
\itemize{
 \item \code{eem_dir_read} will return an object of class \code{eemlist} containing a list of \code{eem}. See \link[eemR]{eem_read} for more details on class structure.
 \item \code{abs_dir_read} will return an object of class \code{abslist} containing a list of \code{abs}. See \link[eemanalyzeR]{abs_read} for more details on class structure.
}
}
\description{
A flexible function that will read all EEM's or absorbance files in a directory even when
it contains other files.
}
\note{
\code{eem_dir_read} is a wrapper for \link[eemR]{eem_read} to read all EEMs in a directory even when
it contains other files.
}
\examples{
#EEM's
# Load all data from directory --------------
eemlist <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"))
# Load all data that matches a pattern from directory -------------
eemlist <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern="SEM")

#Absorbance
#load samples by using pattern to identify absorbance samples -------------
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "abs|ABS")
#load samples by using skip to exclude EEM's and other samples -------------
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), skip = "SEM|BEM|waterfall")

}
