% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir-read.R
\name{dir_read}
\alias{dir_read}
\alias{eem_dir_read}
\alias{abs_dir_read}
\title{Read optics data from directory}
\usage{
eem_dir_read(
  input_dir,
  pattern = NULL,
  skip = "(?i)abs",
  file_ext = "dat",
  recursive = FALSE,
  import_function = "aqualog"
)

abs_dir_read(
  input_dir,
  pattern = NULL,
  skip = "SEM|BEM|Waterfall",
  file_ext = "dat",
  recursive = FALSE
)
}
\arguments{
\item{input_dir}{path to folder containing raw EEMs and/or absorbance files}

\item{pattern}{optional. a character string containing a \code{\link[base]{regular expression}}
to be matched to the files in input_dir.
only files matching the pattern will be loaded.}

\item{skip}{a character string containing a \code{\link[base]{regular expression}} to be
matched to the files in input_dir.
any files matching this string will be ignored.}

\item{file_ext}{character. the file extension of the EEMs or absorbance}

\item{recursive}{logical. should the function recurse into directories?}

\item{import_function}{character or a user-defined function to import an EEM.
for more details see \link[eemR]{eem_read}}
}
\value{
\code{eem_dir_read} returns an object of class eemlist,  containing a list of \code{eem}.
For more details see \link[eemR]{eem}

\code{abs_dir_read} returns an object of class abslist, containing a list of \code{abs}.
For more details see \link[eemanalyzeR]{abs_read}
}
\description{
Wrappers for \link[eemR]{eem_read} and \link[eemanalyzeR]{abs_read} to read all EEMs or absorbance
in a directory into R, even when it contains other files.
}
\examples{
#load EEM's data
# Load all data from directory --------------
eem_list <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"))

# Load all data that matches a pattern from directory -------------
eem_list <- eem_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern="SEM")

#load absorbance data
#load samples by using pattern to identify absorbance samples -------------
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), pattern = "abs|ABS")

#load samples by using skip to exclude EEM's and other samples -------------
abs <- abs_dir_read(system.file("extdata", package = "eemanalyzeR"), skip = "SEM|BEM|waterfall")

}
