% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem-normalize.R
\name{eem_normalize}
\alias{eem_normalize}
\title{Normalize an \code{eem} or \code{eemlist} based on a normalization factor}
\usage{
eem_normalize(eem, factor = NULL)
}
\arguments{
\item{eem}{The \code{eem} or \code{eemlist} to normalize.}

\item{factor}{The normalization factor, either a single value or a vector
of factors. If \code{NULL}, it will normalize to the maximum value for each \code{eem}.}
}
\value{
An \code{eem} or \code{eemlist} where \code{x} has been normalized.
}
\description{
Useful for Raman normalization or normalizing to a maximum of one for blank comparisons.
}
\examples{
# Normalize a single EEM
eem_normal <- eem_normalize(example_eems[1])

# Normalize an entire EEM list
eems_normal <- eem_normalize(example_eems)
}
