% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemanalyzeR-indices.R
\name{eemanalyzeR_indices}
\alias{eemanalyzeR_indices}
\title{Default package methods for fluorescence and absorbance indices}
\usage{
eemanalyzeR_indices(eemlist, abslist, cuvle = 1, qaqc_dir = NULL)
}
\arguments{
\item{eemlist}{An \code{eemlist} object.}

\item{abslist}{An \code{abslist} object.}

\item{cuvle}{Cuvette (path) length in cm.}

\item{qaqc_dir}{File path to the QAQC files generated with \code{\link[=create_mdl]{create_mdl()}} and \code{\link[=create_std]{create_std()}}.
Default is a user-specific data directory \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}.}
}
\value{
A list with two elements:
\itemize{
\item \strong{eem_index}: a \code{data.frame} of all fluorescence indices. Each row corresponds to a single index for a sample.
\item \strong{abs_index}: a \code{data.frame} of all absorbance indices. Each row corresponds to a single index for a sample.
}

Each \code{data.frame} contains the following columns:
\itemize{
\item \strong{sample_name}: name of the sample from the EEM or absorbance list
\item \strong{meta_name}: sample name from metadata if provided; otherwise same as \code{sample_name}
\item \strong{index}: name of the index
\item \strong{value}: calculated value of the index
}
}
\description{
Calculates commonly used absorbance and fluorescence optical indices from \code{eemlist} and \code{abslist}.
For detailed descriptions and references for indices, see the vignette
\href{../doc/eemanalyzeR-indices.html}{\code{vignette("eemanalyzeR-indices")}}.
}
\note{
\itemize{
\item If absorbance is not at a 1 nm interval, it will be interpolated using \code{\link[zoo:na.approx]{zoo::na.approx()}}, which fills in missing values using linear interpolation.
\item If EEM data is not at a 1 nm interval, fluorescence will be interpolated using \code{\link[pracma:interp2]{pracma::interp2()}}.
}
}
\examples{
indices <- eemanalyzeR_indices(
  example_processed_eems,
  example_processed_abs,
  qaqc_dir = system.file("extdata", package = "eemanalyzeR")
)
}
