% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemanalyzeR-indices.R
\name{eemanalyzeR_indices}
\alias{eemanalyzeR_indices}
\title{Default package methods for fluorescence and absorbance indices}
\usage{
eemanalyzeR_indices(eemlist, abslist, cuvle = 1, mdl_dir = .qaqc_dir())
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{cuvle}{cuvette (path) length in cm}

\item{mdl_dir}{file path to the mdl files generated with \link[eemanalyzeR]{get_mdl},
default is a user-specific data directory (\link[rappdirs]{user_data_dir})}
}
\value{
a list with two objects:
\itemize{
  \item eem_index: a data.frame of all the fluorescence indices
  \item abs_index: a data.frame of all the absorbance indices
}
Each data.frame will have four columns:
\itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported, see \href{../doc/eemanalyzeR-indices.html}{vignette}for more information.
 \item value: the value of the index
}
}
\description{
Calculates commonly used absorbance and fluorescence optical indices from eemlist and abslist.
For detailed descriptions and references for indices, see the vignette
\href{../doc/eemanalyzeR-indices.html}{\code{eemanalyzeR-indices}}.
}
\note{
If absorbance is not at a 1 nanometer interval, absorbance will be interpolated using
\link[zoo]{na.approx} which fills in missing values
using linear interpolation.

If EEMs data is not a 1 nanometer interval, fluorescence will be interpolated
using \link[pracma]{interp2}.
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_metadata(metadata, example_eems)
eemlist <- add_blanks(eemlist, validate=FALSE)
eemlist <- process_eem(eemlist, abslist)
indices <- eemanalyzeR_indices(eemlist, abslist,
mdl_dir = system.file("extdata", package = "eemanalyzeR"))
}
