% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemR-indices.R
\name{eemR_indices}
\alias{eemR_indices}
\title{eemR and staRdom methods for fluorescence and absorbance indices}
\usage{
eemR_indices(eemlist, abslist, cuvle = 1)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{cuvle}{cuvette (path) length in cm}
}
\value{
a list with two objects:
\itemize{
  \item eem_index: a data.frame of all the fluorescence indices
  \item abs_index: a data.frame of all the absorbance indices
}
Each data.frame will have four columns:
\itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported, see details for more information.
 \item value: the value of the index
}
}
\description{
Calculates commonly used absorbance and fluorescence optical indices from eemlist and abslist using
\href{https://CRAN.R-project.org/package=eemR}{eemR} and \href{https://CRAN.R-project.org/package=staRdom}{staRdom} functions.
}
\details{
\strong{Absorbance Indices}: Absorbance indices (a254, a300, E2_E3, E4_E6, S275_295, S350_400, S300_700, SR)
are generated using the staRdom function \link[staRdom]{abs_parms}.

\strong{Fluorescence Indices}: Fluorescence indices are generated using the eemR functions:
\itemize{
\item Coble peaks (b, t, a, m, c): \link[eemR]{eem_coble_peaks}
\item fi: \link[eemR]{eem_fluorescence_index}
\item hix: \link[eemR]{eem_humification_index}, where hix_scaled uses scale=TRUE
\item bix: \link[eemR]{eem_biological_index}
}
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_metadata(metadata, example_eems)
indices <- eemR_indices(eemlist, abslist)
}
