% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_abs}
\alias{example_abs}
\title{Down-scaled absorbance data}
\format{
An \code{abslist} with three samples
\itemize{
  \item{\strong{B1S1ExampleBlankABS}: Absorbance measured for a sample of Type 1 water}
  \item{\strong{B1S2ExampleTeaStdABS}: Absorbance measured for a sample of dilute unsweetened tea}
  \item{\strong{B1S3ExampleSampleABS}: Absorbance measured for a sample of stream water from Oregon, USA}
}
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022-11-14)
}
\usage{
example_abs
}
\description{
A subset of water samples run on a Horiba Aqualog for absorbance. This data only includes
the absorbance data associated with the samples in \link{example_eems}. Data was loaded using the
\link[eemanalyzeR]{abs_dir_read} function. For more details on the format of an eemlist, see \link[eemanalyzeR]{abs_read}.
}
\keyword{datasets}
