% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_blanks}
\alias{example_blanks}
\title{Down-scaled excitation emission matrices of sample blanks}
\format{
An eemlist with three samples:
\describe{
  \item{ExampleBlank}{Fluoresence measured for a blank of blank Milliq water}
  \item{ExampleTeaStd}{Fluoresence measured for a blank of blank Milliq water}
  \item{ExampleSample}{Fluoresence measured for a blank of blank Milliq water}
}
}
\source{
Oregon State University Forest Watershed Hydrology Lab
}
\usage{
example_blanks
}
\description{
A subset of water samples run on a Horiba Aqualog for excitation emission matrices (EEMs). This data only includes
the blanks associated with the samples in \link{example_samples}. Data was loaded using \link[eemR]{eem_read}.
For more details on the format of an eemlist, see the \link[eemR]{eemR} package.
}
\keyword{datasets}
