% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_processed_abs}
\alias{example_processed_abs}
\title{Down-scaled and processed absorbance data}
\format{
An `abslist` with four samples:
- **B1S1ExampleBlankABS**: Absorbance measured for a sample of Type 1 water
- **B1S2ExampleTeaStdABS**: Absorbance measured for a sample of dilute unsweetened tea
- **B1S3ExampleSampleABS**: Absorbance measured for a sample of stream water from Oregon, USA
- **ManualExampleTeaAbsSpectraGraphs**: Absorbance measured for a sample of dilute unsweetened tea
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022-11-14)
}
\usage{
example_processed_abs
}
\description{
Processed version of [example_abs()] using `eemanalyzeR`.
}
\keyword{datasets}
