% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_processed_eems}
\alias{example_processed_eems}
\title{Down-scaled and processed excitation emission matrices}
\format{
An `eemlist` with four samples:
- **B1S1ExampleBlankSEM**: Fluorescence measured for a sample of Type 1 water
- **B1S2ExampleTeaStdSEM**: Fluorescence measured for a sample of dilute unsweetened tea
- **B1S3ExampleSampleSEM**: Fluorescence measured for a sample of stream water from Oregon, USA
- **ManualExampleTeaWaterfallPlotSample**: Fluorescence measured for a sample of dilute unsweetened tea
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022-11-14)
}
\usage{
example_processed_eems
}
\description{
Processed version of [example_eems()] using `eemanalyzeR`.
}
\keyword{datasets}
