% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_samples}
\alias{example_samples}
\title{Down-scaled excitation emission matrices of samples}
\format{
An eemlist with three samples:
\describe{
  \item{ExampleBlank}{Fluoresence measured for a sample of blank Milliq water}
  \item{ExampleTeaStd}{Fluoresence measured for a sample of dillute unsweetened tea}
  \item{ExampleSample}{Fluoresence measures for a sample of stream water from Oregon, USA}
}
}
\source{
Oregon State University Forest Watershed Hydrology Lab
}
\usage{
example_samples
}
\description{
A subset of water samples run on a Horiba Aqualog for excitation emission matrices (EEMs). This data only includes
the samples associated with the blanks in \link{example_blanks}. Data was loaded using \link[eemR]{eem_read}.
For more details on the format of an eemlist, see the \link[eemR]{eemR} package.
}
\keyword{datasets}
