% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-data.R
\name{export_data}
\alias{export_data}
\title{Export Processed Optics Data}
\usage{
export_data(
  eemlist,
  abslist,
  filename,
  output_dir = NULL,
  meta = NULL,
  indices = NULL,
  eem_plot = NULL,
  abs_plot = NULL,
  csv = FALSE
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{filename}{a character, creates a folder with this name within the \code{output_dir} and uses this name for saving file names}

\item{output_dir}{the location to save the data to. if no location is specified it will save to the temporary directory}

\item{meta}{optional, a \code{data.frame} of metadata}

\item{indices}{optional, if \code{NULL} indices will not be exported. If a list of indices are included
these will be exported}

\item{eem_plot}{optional, if \code{NULL} EEM's plots will not be exported. If a list of plots are included
these will be exported as .png files. See \link[=plot.eemlist]{plot.eemlist} for plotting.}

\item{abs_plot}{optional, if \code{NULL} absorbance plots will not be exported. If a plot is included
these will be exported as a .png file. See \link[=plot.abslist]{plot.abslist} for plotting.}

\item{csv}{logical, if TRUE processed EEMs and absorbance data will be written to the \code{output_dir} as .csv files}
}
\value{
A list of the processed data which is also saved to the output directory specified.
The list contains the following items:
\itemize{
\item \strong{eemlist:} the \code{eemlist}
\item \strong{abslist:} the \code{abslist}
\item \strong{readme:} a character vector containing information about the data processing steps
\item \strong{meta:} the metadata associated with the samples, may be \code{NULL} if not provided
\item \strong{indices:} a list of EEMs and absorbance indices, may be \code{NULL} if not provided
\item \strong{eem_plot:} a list of EEMs plots, may be \code{NULL} if not provided
\item \strong{abs_plot:} a ggplot2 object of the absorbance data, may be \code{NULL} if not provided
}
}
\description{
Export Processed Optics Data
}
\examples{
eem_plots <- plot(example_processed_eems)
abs_plot <- plot(example_processed_abs)

indices <- get_indices(example_processed_eems, example_processed_abs)
data <- export_data(example_processed_eems, example_processed_abs,
    filename="eemanalyzeR_example",indices = indices,
    eem_plot=eem_plots, abs_plot = abs_plot, meta=metadata)
}
