% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-index.R
\name{format_index}
\alias{format_index}
\title{Format Index and Flag Data}
\usage{
format_index(x, index, value, flag)
}
\arguments{
\item{x}{an \code{eemlist} or \code{abslist} object}

\item{index}{a character, the name of the index}

\item{value}{the index values, use \link[eemanalyzeR]{get_fluorescence} or \link[eemanalyzeR]{get_absorbance} to get index values}

\item{flag}{the flag values, must be the same length as value, use \link[eemanalyzeR]{flag_missing} to generate flags}
}
\value{
a data.frame with four columns:
#' \itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported
 \item value: the value of the index
}
}
\description{
Takes index values, QA/QC flags, and dataset and combines and combines data and formats data for further processing.
Will remove any values with a flag of MDL01 as these are below the detection limit and won't be reported.
}
\examples{
 ex <- 240:260
 em <- 300:320
 vals <- get_fluorescence(example_eems, ex, em, stat = "max")
 flags <- flag_missing(example_eems, ex=ex, em=em, all=FALSE)
 index_formatted <- format_index(example_eems, "test_index", vals, flags)
}
