% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-abs-slope.R
\name{get_abs_slope}
\alias{get_abs_slope}
\title{Calculate spectral slopes}
\usage{
get_abs_slope(abs, lim, cuvle = 1)
}
\arguments{
\item{abs}{an object of class \code{abs} or \code{abslist}}

\item{lim}{a vector of length two containing lower and upper limits for wavelengths to use for calculating slope}

\item{cuvle}{cuvette (path) length in cm}
}
\value{
A vector of spectral slope values, one for each sample in abslist. If a value cannot be extracted, DATA04 will be returned.
}
\description{
A wrapper for \link[staRdom]{abs_fit_slope} which interpolates the data as needed, calculates absorption, calculates the slope, and flags any missing values.
}
\examples{
S275_295 <- get_abs_slope(example_absorbance, lim=c(275,295))
}
