% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-absorbance.R
\name{get_absorbance}
\alias{get_absorbance}
\title{Extract absorbance at a given wavelength}
\usage{
get_absorbance(abs, wl, cuvle = 1, suva = FALSE)
}
\arguments{
\item{abs}{An `abs` or `abslist` object.}

\item{wl}{A vector of absorbance wavelengths.}

\item{cuvle}{Cuvette (path) length in cm.}

\item{suva}{If `TRUE`, returns specific absorbance (SUVA).}
}
\value{
A vector of absorbance (or specific absorbance) values.
If a value cannot be extracted, "DOC01" is returned.
}
\description{
Returns the absorbance value at a specified wavelength, reported in
\ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^-1}}.
Optionally calculates specific absorbance (SUVA), reported in
\ifelse{html}{\out{m<sup>-1</sup>}}{\eqn{m^-1}}.
}
\details{
If metadata is not present, SUVA cannot be calculated and `NA` will be returned.
To add metadata, use [add_metadata()].

If the wavelength requested is not
in the absorbance data, it will be interpolated first using
[abs_interp()].
}
\examples{
abslist <- add_metadata(metadata, example_abs)

a254 <- get_absorbance(abslist, 254)
suva254 <- get_absorbance(abslist, 254, suva = TRUE)
}
