% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fluorescence.R
\name{get_fluorescence}
\alias{get_fluorescence}
\title{Get fluorescence within a specified range}
\usage{
get_fluorescence(eem, ex, em, stat = "max")
}
\arguments{
\item{eem}{an \code{eem} or \code{eemlist} object containing EEM's data}

\item{ex}{a vector of excitation wavelengths}

\item{em}{a vector of emission wavelengths}

\item{stat}{either "max" to get the maximum value within the specified range or "sum" to get the sum across the specified range}
}
\value{
A vector of fluorescence values. If a value cannot be extracted, NA will be returned.
}
\description{
Gets the fluorescence within a range of excitation and emission wavelengths and returns the maximum or sum within that range.
Missing values are interpolated using \link[pracma]{interp2}.
}
\examples{
pA <- get_fluorescence(example_eems, ex=250:260, em=380:480)
pD <- get_fluorescence(example_eems, ex=390, em=509)
pA_sum <- get_fluorescence(example_eems, ex=250:260, em=380:480, stat="sum")
}
