% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indices.R
\name{get_indices}
\alias{get_indices}
\title{Get fluorescence and absorbance indices}
\usage{
get_indices(
  eemlist,
  abslist,
  index_method = "eemanalyzeR",
  return = "long",
  cuvle = 1,
  mdl_dir = .qaqc_dir(),
  arg_names = NULL
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{index_method}{currently supports "eemanalyzeR", "eemR", and "usgs". See details for more information.}

\item{return}{either "long" or "wide" to specify the format of the indices data.frames}

\item{cuvle}{cuvette (path) length in cm}

\item{mdl_dir}{file path to the mdl files generated with \link[eemanalyzeR]{create_mdl} and \link[eemanalyzeR]{create_tea_std},
default is a user-specific data directory (\link[rappdirs]{user_data_dir})}

\item{arg_names}{optional argument used to pass arguments from higher level functions for writing the readme.}
}
\value{
a list with two objects:
\itemize{
  \item eem_index: a data.frame of all the fluorescence indices
  \item abs_index: a data.frame of all the absorbance indices
}
If \code{return} is 'long' each data.frame will have four columns:
\itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported, see details for more information.
 \item value: the value of the index
 \item QAQC_flag: any flags associated with the sample, see details for more information.
}

If \code{return} is 'wide' each data.frame will have a row for each sample, where the columns are the different indices. In this format QAQC flags
are indicated in the individuals cells, either by itself if no value was able to be reported, or combined with the value where applicable to have the form
'value_flag'.
}
\description{
Calculates commonly use indices from absorbance and excitation emission matrix (EEM) data. Also
checks and flags the indices for potential errors or issues.
}
\details{
\strong{Index Methods}

This function allows for three different sets of indices:
\href{../doc/eemanalyzeR-indices.html}{\code{eemanalyzeR}} (default),
eemR and usgs. See the links for more information on the indices in each method and how they're calculated.
Custom functions can also be used to calculate indices, see vignette for more information on how to do this.

\strong{QA/QC Flags}

The index values will be checked for potential errors prior to reporting and flagged if necessary:
\itemize{
 \item DATA01: Missing data required to calculate the index
 \item DATA02: Missing some wavelengths required to calculate the index, value may be inaccurate
 \item DATA03: Unable to calculate ratio because denominator was zero
 \item DATA04: Spectral slope was unable to be calculated
 \item DOC01: Missing dissolved organic carbon data, so index was not able to be calculated
 \item MDL01: All values were below the MDL for those wavelengths
 \item MDL02: One of more of the values were below the MDL for those wavelengths, use data cautiously
 \item MDL03: Index was a ratio, and either numerator or denominator was completely below the MDL
 \item NEG01: Value was negative
 \item NOISE01: Value was below signal to noise ratio and therefore was not calculated
 \item NOISE02: Value was not able to be checked for signal to noise ratio
 \item STD01: The tea check standard values were outside of the tolerance threshold for this index
 \item VAL01: Value was below expected values for this index, normal ranges can vary by sample matrix, but please check value for accuracy
 \item VAL02: Value was above expected values for this index, normal ranges can vary by sample matrix, but please check value for accuracy
}
}
\examples{
indices <- get_indices(example_processed_eems, example_processed_abs,
mdl_dir = system.file("extdata", package = "eemanalyzeR"))
}
