% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indices.R
\name{get_indices}
\alias{get_indices}
\title{Get fluorescence and absorbance indices}
\usage{
get_indices(
  eemlist,
  abslist,
  index_method = "eemanalyzeR",
  return = "long",
  cuvle = 1
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{index_method}{currently supports "eemanalyzeR", "eemR", and "usgs". See details for more information.}

\item{return}{either "long" or "wide" to specify the format of the indices data.frames}

\item{cuvle}{cuvette (path) length in cm}
}
\value{
a list with two objects:
\itemize{
  \item eem_index: a data.frame of all the fluorescence indices
  \item abs_index: a data.frame of all the absorbance indices
}
Each data.frame will have four columns:
\itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported, see details for more information.
 \item value: the value of the index
 \item QAQC_flag: any flags associated with the sample, see details for more information.
}
}
\description{
Get fluorescence and absorbance indices
}
\details{
\strong{Index Methods}

This function allows for three different sets of indices:
\href{../doc/eemanalyzeR-indices.html}{\code{eemanalyzeR}} (default),
eemR and usgs. See the links for more information on the indices in each method and how they're calculated.
Custom functions can also be used to calculate indices, see vignette for more information on how to do this.

\strong{QA/QC Flags}

The index values will be checked for potential errors prior to reporting and flagged if necessary:
\itemize{
 \item DATA_01: Missing data required to calculate the index
 \item DATA_02: Missing some wavelengths required to calculate the index, value may be inaccurate
 \item DOC_01: Missing dissolved organic carbon data, so index was not able to be calculated
 \item NEG_01: Value was negative
 \item NOISE_01: Value was below signal to noise ratio and therefore was not calculated
 \item NOISE_02: Value was below signal to noise ratio and may be inaccurate
 \item VAL_01: Value was below expected values for this index, please check for accuracy
 \item VAL_02: Value was above expected values for this index, please check for accuracy
}
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_metadata(metadata, example_eems)
indices <- get_indices(eemlist, abslist)
}
