% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indices-function.R
\name{get_indices_function}
\alias{get_indices_function}
\title{Get function to generate fluorescence and absorbance indices}
\usage{
get_indices_function(index_method = "eemanalyzeR")
}
\arguments{
\item{index_method}{Either "eemanalyzeR", "eemR", "usgs", or a custom function.}
}
\value{
A function used to generate indices.
}
\description{
Returns either a user-supplied custom function or one of the preset index
generators.
}
\details{
Preset methods correspond to the following functions:
\itemize{
\item \strong{eemanalyzeR}: \code{\link[=eemanalyzeR_indices]{eemanalyzeR_indices()}}
\item \strong{eemR}: \code{\link[=eemR_indices]{eemR_indices()}}
\item \strong{usgs}: \code{\link[=usgs_indices]{usgs_indices()}}
}
}
