% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indices-function.R
\name{get_indices_function}
\alias{get_indices_function}
\title{Get function to generate fluorescence and absorbance indices}
\usage{
get_indices_function(index_method = "eemanalyzeR")
}
\arguments{
\item{index_method}{currently supports "eemanalyzeR", "eemR", and "usgs" or a custom function}
}
\value{
a function used to generate indices
}
\description{
Accepts either a custom function to generate indices (see vignette) or uses preset methods.
}
\details{
For more details on the preset methods see the following functions:
\itemize{
\item eemanalyzeR: \link[eemanalyzeR]{eemanalyzeR_indices}
\item eemR: \link[eemanalyzeR]{eemR_indices}
\item usgs: \link[eemanalyzeR]{usgs_indices}
}
}
