% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_qaqc}
\alias{get_qaqc}
\title{Look for MDL files}
\usage{
get_qaqc(qaqc_dir, type, method = get_qaqc_method(), quiet = FALSE)
}
\arguments{
\item{qaqc_dir}{file path to the mdl files generated with \link[eemanalyzeR]{create_mdl}}

\item{type}{Either "mdl" or "check-std" to specify the type of QA/QC files to return.}

\item{method}{Character of the method to use for the QAQC files.}

\item{quiet}{Logical. Should function warn if default is used?}
}
\description{
If they exist will load, if not will warn. Writes the appropriate message about
MDL in the readme.
}
\examples{
#No directory will return NULL
get_qaqc(NA, type="mdl", quiet =TRUE)

#Otherwise will try to return the requested QAQC files
mdl <- get_qaqc(file.path(system.file("extdata", package = "eemanalyzeR")), type = "mdl")
plot(mdl$eem_mdl)
}
