% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qaqc.R
\name{get_qaqc}
\alias{get_qaqc}
\title{Look for and load QAQC files}
\usage{
get_qaqc(qaqc_dir, type, method = get_qaqc_method(), quiet = TRUE)
}
\arguments{
\item{qaqc_dir}{file path to the mdl files generated with \link[eemanalyzeR]{create_mdl}}

\item{type}{Either "mdl" or "check-std" to specify the type of QA/QC files to return.}

\item{method}{Character of the method to use for the QAQC files.}

\item{quiet}{Logical. Should function warn if default is used?}
}
\value{
A list of length 3:
\enumerate{
\item eem-*: \code{NULL} if not found, otherwise the requested QAQC file for eems.
\item abs-*: \code{NULL} if not found, otherwise the requested QAQC file for absorbance.
\item method: The name of the method chosen, \code{NULL} if no files are found.
}
}
\description{
Looks for the specified QAQC files (MDL and check standards) within the \code{qaqc_dir} specified in
the config file. If more than one set is found, it will prompt the user to specify which
method they want to use. This will be stored in the local copy of the config which will
be maintained for the session. Will also provide warnings and write to the \code{readme} to note if
no files were found or what method was used.
}
\examples{
#No directory will return NULL
get_qaqc(NA, type="mdl", quiet =TRUE)

#Otherwise will try to return the requested QAQC files
mdl <- get_qaqc(file.path(system.file("extdata", package = "eemanalyzeR")), type = "mdl")
plot(mdl$eem_mdl)
}
