% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample-info.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Extract info from eemlist or abslist}
\usage{
get_sample_info(x, info)
}
\arguments{
\item{x}{an object of class:
\itemize{
\item{\code{eemlist}}
\item{\code{eem}}
\item{\code{abslist}}
\item{\code{abs}}
}}

\item{info}{the name of the component within the \code{eem} or \code{abs} to extract.
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{abs_read} for base \code{abs} names.
see \link[eemanalyzeR]{add_metadata} for extended \code{eem} component names.}
}
\value{
if \code{x} is an \code{eemlist} or \code{abslist}:

if requested component (\code{info}) is a:
 \itemize{
 \item vector of length one: a vector of values where each value corresponds to the info for each sample
 \item vector of length greater than one: a matrix of values where each row corresponds to the info for each sample
 \item matrix: returns a list of matrices where each list item corresponds to the info for each sample
 }

if \code{x} is an \code{eem} or \code{abs} it will return the component, maintaining class and structure.
}
\description{
Helper function that build upon the \link[eemR]{eem_names}
function. Extracts components of an \code{eemlist/eem} and \code{abslist/abs} object.
}
\examples{
#get names
get_sample_info(example_eems, "sample")
get_sample_info(example_abs, "sample")

#get analysis_date
eemlist <- add_metadata(metadata, example_eems)
get_sample_info(eemlist, "analysis_date")

#get doc for fifth eem
eemlist <- add_metadata(metadata, example_eems)
get_sample_info(eemlist[[5]], "doc_mgL")

#get emission wavelengths
get_sample_info(example_eems, "em")

#get absorbance data
get_sample_info(example_abs, "data")
}
