% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ife_correct.R
\name{ife_correct}
\alias{ife_correct}
\title{Perform Inner-Filter Corrections}
\source{
Massicotte P (2019). eemR: Tools for Pre-Processing
Emission-Excitation-Matrix (EEM) Fluorescence Data. R package
version 1.0.1, \link[https://CRAN.R-project.org/package=eemR]{https://CRAN.R-project.org/package=eemR}
}
\usage{
ife_correct(eemlist, abslist, pathlength = 1)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{pathlength}{a numeric value indicating the pathlength (in cm) of the
cuvette used for absorbance measurement. Default is 1 (1 cm).}
}
\value{
an object of class \code{eemlist}
}
\description{
A wrapper for eemR function \link[eemR]{eem_inner_filter_effect},
modified to work with the augmented eemlist structure and automatically
cut EEM's wavelengths down to match absorbance data instead of returning an error.
}
\details{
The function links the absorbance and EEM's data by the metadata
name which should be the same between the two datasets, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.
}
\examples{
eemlist <- add_metadata(metadata,example_eems)
abslist <- add_metadata(metadata, example_absorbance)
correct_eem <- ife_correct(eemlist, abslist)
}
