% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-check.R
\name{meta_check}
\alias{meta_check}
\title{Run tests to validate metadata}
\usage{
meta_check(meta)
}
\arguments{
\item{meta}{A \code{data.frame} containing sample metadata.}
}
\value{
A corrected metadata \code{data.frame} with formatting issues resolved where
possible.
}
\description{
Runs a series of checks to identify missing or malformed metadata and
corrects issues when possible. Ensures that column formats conform to
expectations.
}
\examples{
metadata <- meta_read(system.file("extdata", package = "eemanalyzeR"))
metadata <- eemanalyzeR:::meta_check(metadata)

}
\seealso{
\code{\link[=meta_read]{meta_read()}}
}
