% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{meta_check}
\alias{meta_check}
\title{Run tests to validate metadata}
\usage{
meta_check(meta)
}
\arguments{
\item{meta}{a data.frame with metadata}
}
\value{
a \code{data.frame} (meta), with errors fixed
}
\description{
Run tests to validate metadata
}
\examples{
metadata <- meta_read(system.file("extdata", package = "eemanalyzeR"))
metadata <- eemanalyzeR:::meta_check(metadata)
}
