% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-read.R
\name{meta_read}
\alias{meta_read}
\title{Read metadata file from directory}
\usage{
meta_read(input, sheet = NULL, validate = TRUE)
}
\arguments{
\item{input}{path to folder containing metadata file, or name of metadata file}

\item{sheet}{name of sheet containing metadata. only required if metadata isn't the first sheet of xlsx file}

\item{validate}{logical. should metadata structure be checked for potential issues that will cause issues during
further processing? highly recommended to keep as TRUE.}
}
\value{
a data.frame containing metadata of samples
}
\description{
Reads a .csv or .xlsx file containing metadata related to absorbance and EEM's data.
Please see \link[eemanalyzeR]{metadata} for the required columns and structure.
}
\examples{
metadata <- meta_read(system.file("extdata", package = "eemanalyzeR"))
}
