% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{meta_read}
\alias{meta_read}
\title{Read metadata file from directory}
\usage{
meta_read(input_dir, name = NULL, sheet = NULL, validate = T)
}
\arguments{
\item{input_dir}{path to folder containing metadata file}

\item{name}{name of metadata file. optional if metadata is only xlsx or csv file in input_dir
if not specified function will attempt to load any xlsx or csv file in directory and return an error if there is more than one}

\item{sheet}{name of sheet containing metadata. only required if metadata isn't the first sheet}

\item{validate}{logical. should metadata structure be checked for potential issues that will cause issues during
further processing? highly recommended to keep as TRUE.}
}
\value{
a data.frame containing metadata of samples
}
\description{
Reads a .csv or .xlsx file containing metadata related to absorbance and EEM's data.
Please see \link[eemanalyzeR]{metadata} for the required columns and structure.
}
\examples{
metadata <- meta_read(system.file("extdata", package = "eemanalyzeR"))
}
