% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Example metadata}
\format{
A data.frame with 3 rows and 11 columns:
\describe{
  \item{index}{Simply the numeric order of entry (e.g. 1, 2, 3, 4, etc…)}
  \item{analysis_date}{optional: Date samples were ran on instrument (not collected in the field)}
  \item{description}{optional: A brief description of the sample collected}
  \item{data_identifier}{REQUIRED. The file name from the aqualog, needs to match exactly}
  \item{replicate_no}{REQUIRED. A number indicating if the sample was replicated, an unreplicated sample should get a 1. Analytical replicates must have identical sample names.}
  \item{integration_time_s}{REQUIRED. Integration time of sample (e.g. 1, 2, etc…)}
  \item{dilution}{REQUIRED. Include dilution factor here as decimal format (e.g. a 2-fold dilution with 1 part sample and 1 part water will have a dilution factor of 0.5). If no dilution, put 1}
  \item{RSU_area_1s}{REQUIRED: The RSU Adjust Area from the Raman Test, used to normalize the data from day to day}
  \item{is_blank}{optional: Flag for whether the sample is a blank sample (e.g. MilliQ water). Possible value of TRUE or FALSE}
  \item{is_check}{optional:Flag for whether the sample is a check standard (usually SRM Tea). Possible value of TRUE or FALSE}
  \item{run_type}{REQUIRED. How the samples were run (manual or sampleQ)}
  \item{collect_date}{optional: The date the water samples were collected}
  \item{DOC_mg_L}{optional: DOC concentration in mg/L of the original sample (not the diluted sample)}
  \item{Notes}{optional: Any notes from sample collection or sample running, will not be used in analysis}
}
}
\source{
Oregon State University Forest Ecohydrology and Watershed Sciences Lab (2022-11-14)
}
\usage{
metadata
}
\description{
Metadata associated with the samples in \link{example_eems}. Provides an example of the structure
and requirements for metadata needed to run samples.
}
\keyword{datasets}
