% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.eem}
\alias{plot.eemlist}
\alias{plot.abs}
\alias{plot.abslist}
\title{Default plot methods for optical data with ggplot2}
\usage{
\method{plot}{eem}(
  x,
  nbin = 8,
  equal_scale = FALSE,
  pal = NULL,
  title = "none",
  remove_lower = FALSE,
  annotate = FALSE,
  index_method = "eemanalyzeR",
  ...
)

\method{plot}{eemlist}(
  x,
  nbin = 8,
  equal_scale = FALSE,
  pal = NULL,
  remove_lower = FALSE,
  title = "none",
  annotate = FALSE,
  index_method = "eemanalyzeR",
  ...
)

\method{plot}{abs}(x, pal = NULL, ...)

\method{plot}{abslist}(x, pal = NULL, ...)
}
\arguments{
\item{x}{An \code{eemlist}, \code{eem}, \code{abslist}, or \code{abs} object.}

\item{nbin}{Number of bins used in the contour plot.}

\item{equal_scale}{Logical. If \code{TRUE}, sets the scale the same for all
plots in an \code{eemlist}.}

\item{pal}{Color palette for the fill scale. Defaults to \code{\link[pals:continuous]{pals::parula()}}.
If fewer colors are provided than required, \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}
is used to fill in colors.}

\item{title}{Either "none", "meta_name", or "sample" which indicates what to use for the plot title.}

\item{remove_lower}{Logical. If \code{TRUE}, sets values below the first-order
Rayleigh line to \code{NA}, which can reduce artifacts affecting the color scale.}

\item{annotate}{Logical. If \code{TRUE}, displays index regions on EEM plots.}

\item{index_method}{Either "eemanalyzeR", "eemR", "usgs".}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
\itemize{
\item If \code{x} is an \code{eem}, \code{abs}, or \code{abslist}: a single \code{ggplot2} object.
\item If \code{x} is an \code{eemlist}: a list of \code{ggplot2} objects.
}
}
\description{
Used to make a nice plot of one or multiple excitation emission matrices
(EEMs) or absorbance spectra using ggplot2.
}
\details{
Single EEM plots return a \code{ggplot2} object, compatible with other \code{ggplot2}
modifications. Multiple EEMs return a list of \code{ggplot2} objects, which can
be individually modified. See examples for usage.
}
\note{
Slow plotting may be due to your default graphics device. See
\url{https://forum.posit.co/t/graphics-not-working-in-rstudio/149111} for guidance.
}
\examples{
eems <- example_processed_eems
abs <- example_processed_abs

# plot just one eem/abs
plot(eems[[3]])
plot(abs[[3]])

# plot all in an eemlist or abslist
eem_plots <- plot(eems)
abs_plots <- plot(abs)

# change color scale
plot(eems, pal = c("darkblue", "lightblue"))

# make color bar consistent across all plots
plot(eems[2:3], equal_scale = TRUE)

# customize using ggplot2 commands
plot(eems[[2]]) + ggplot2::labs(title = "Test EEM")
eem_plots[[3]] + ggplot2::labs(title = "Test EEM")

# modify then arrange together
eem_plots[[3]] <- eem_plots[[3]] + ggplot2::labs(title = "Test EEM")
print(ggpubr::ggarrange(plotlist = eem_plots))

# remove lower area below rayleigh line
plots <- plot(eems[[3]], remove_lower = TRUE)

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
