% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-eem.R
\name{process_eem}
\alias{process_eem}
\title{Process and correct excitation emission matrices}
\usage{
process_eem(
  eemlist,
  abslist,
  ex_clip = c(247, 450),
  em_clip = c(247, 600),
  type = c(TRUE, TRUE, TRUE, TRUE),
  width = c(16, 3, 30, 10),
  interpolate = c(TRUE, TRUE, FALSE, FALSE),
  method = 1,
  cores = 1,
  pathlength = 1
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{ex_clip}{numerical vector of length two with the minimum and maximum wavelengths to clip the excitation wavelengths to.}

\item{em_clip}{numerical vector of length two with the minimum and maximum wavelengths to clip the emission wavelengths to.}

\item{type}{a logical vector of length four indicating which scattering lines should be removed.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.
Set \code{TRUE} for scattering lines that should be removed.}

\item{width}{a numeric vector of length four containing width of
scattering in nanometers to remove. The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{interpolate}{a logical vector of length four indicating which scattering lines to interpolate.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{method}{numeric 0 to 4 for interpolation method to use. Default is 1. See \link[staRdom]{eem_interp} for more info.}

\item{cores}{The number of cores used for parallel computation of interpolation.}

\item{pathlength}{a numeric value indicating the pathlength (in cm) of the
cuvette used for absorbance measurement. Default is 1 (1 cm).}
}
\value{
an object of class \code{eemlist}
}
\description{
A wrapper to combine all the processing functions together to make a cleaner workflow.
Processing steps include:
\itemize{
\item{\link[eemanalyzeR]{subtract_blank}: which subtracts the blanks from the samples}
\item{\link[eemanalyzeR]{remove_scattering}: which removes scattering lines with optional interpolation}
\item{\link[eemanalyzeR]{ife_correct}: which corrects the fluoresence intensity for inner filter effects}
\item{\link[eemanalyzeR]{raman_normalize}: which normalizes the fluoresence intensity to the raman unit area}
\item{\link[eemanalyzeR]{correct_dilution}: which corrects the fluoresence intensity for any dilutions}
\item{\link[eemR]{eem_cut}: which clips the area of the excitation emission matrices to the area of interest}
}
}
\details{
The function links the absorbance and EEM's data by the metadata
name which should be the same between the two datasets, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.
}
\examples{
eemlist <- add_metadata(metadata, example_eems)
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_blanks(eemlist, validate=FALSE)
corrected_eem <- process_eem(eemlist, abslist)
plot_eem(corrected_eem)

}
