% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman-normalize.R
\name{raman_normalize}
\alias{raman_normalize}
\title{Perform Raman Normalization}
\source{
Lawaetz, A. J., & Stedmon, C. A. (2009). Fluorescence Intensity Calibration Using
the Raman Scatter Peak of Water. Applied Spectroscopy, 63(8), 936-940.
\href{https://journals.sagepub.com/doi/10.1366/000370209788964548}{https://journals.sagepub.com/doi/10.1366/000370209788964548}
}
\usage{
raman_normalize(eemlist)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}
}
\value{
an object of class \code{eemlist}
}
\description{
Normalizes the EEM intensity to a standard scale of Raman Units (R.U.) based
on the area of the water Raman peak (Lawaetz and Stedmon 2009).
}
\details{
The function uses the raman area provided by the metadata to normalize the EEMs, because of this,
samples must have metadata already added to the samples using \link[eemanalyzeR]{add_metadata}.
}
\examples{
eemlist <- add_metadata(metadata,example_eems)
eemlist <- subset_qaqc(eemlist)
correct_eem <- raman_normalize(eemlist)
}
