% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-scattering.R
\name{remove_scattering}
\alias{remove_scattering}
\title{Remove scattering lines}
\usage{
remove_scattering(
  eemlist,
  type = c(TRUE, TRUE, TRUE, TRUE),
  width = c(16, 3, 30, 10),
  interpolate = c(TRUE, TRUE, FALSE, FALSE),
  method = 1,
  cores = 1
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data.}

\item{type}{a logical vector of length four indicating which scattering lines should be removed.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.
Set \code{TRUE} for scattering lines that should be removed.}

\item{width}{a numeric vector of length four containing width of
scattering in nanometers to remove. The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{interpolate}{a logical vector of length four indicating which scattering lines to interpolate.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{method}{numeric 0 to 4 for interpolation method to use. Default is 1. See \link[staRdom]{eem_interp} for more info.}

\item{cores}{The number of cores used for parallel computation of interpolation.}
}
\value{
an object of class \code{eemlist}
}
\description{
TODO add info about scattering lines here, add in readme notes
}
\details{
A modified version of the staRdom function \link[staRdom]{eem_rem_scat}, modified to allow
selective interpolation instead of all or nothing.
}
\examples{
#default settings (remove all, interpolate only raman lines)
eemlist <- remove_scattering(example_eems)
staRdom::ggeem(eemlist[[6]])

#interpolate all
eemlist <- remove_scattering(example_eems, interpolate=c(TRUE,TRUE,TRUE,TRUE))
staRdom::ggeem(eemlist[[6]])

#only remove only rayleigh lines
eemlist <- remove_scattering(example_eems, type=c(FALSE,FALSE,TRUE,TRUE))
staRdom::ggeem(eemlist[[6]])

#change the width of the lines
eemlist <- remove_scattering(example_eems, width=c(16,3,100,40))
staRdom::ggeem(eemlist[[6]])
}
