% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-scattering.R
\name{remove_scattering}
\alias{remove_scattering}
\title{Remove Raman and Rayleigh scattering lines}
\usage{
remove_scattering(
  eemlist,
  type = c(TRUE, TRUE, TRUE, TRUE),
  width = c(16, 3, 30, 10),
  interpolate = c(TRUE, TRUE, FALSE, FALSE),
  method = 1,
  cores = 1
)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data.}

\item{type}{a logical vector of length four indicating which scattering lines should be removed.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.
Set \code{TRUE} for scattering lines that should be removed.}

\item{width}{a numeric vector of length four containing width of
scattering in nanometers to remove. The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{interpolate}{a logical vector of length four indicating which scattering lines to interpolate.
The order is "raman1", "raman2", "rayleigh1" and "rayleigh2" scattering.}

\item{method}{numeric 0 to 4 for interpolation method to use. Default is 1. See \link[staRdom]{eem_interp} for more info.}

\item{cores}{The number of cores used for parallel computation of interpolation.}
}
\value{
an object of class \code{eemlist}
}
\description{
Raman and Rayleigh scattering can cause signal interference with the sample fluorescence signal,
thus it is often advisable to remove them. Raman scatting is a form of inelastic scattering,
that is due light interacting with the solvent, shifting to longer wavelengths. Rayleigh scattering is elastic scattering, where light
bounces off molecules in the sample without losing energy, thus this line will always appear on the 1:1 line
and tends to be quite intense. Second order Raman and Rayleigh scattering may also be observed,
which is due to the interaction between two photons of the same energy which doubles to excitation
wavelength.
}
\note{
This function is a modified version of the staRdom function \link[staRdom]{eem_rem_scat}, modified to allow
selective interpolation instead of all or nothing.
}
\examples{
#default settings (remove all, interpolate only raman lines)
eemlist <- remove_scattering(example_eems)
plot_eem(eemlist[[6]])

#interpolate all
eemlist <- remove_scattering(example_eems, interpolate=c(TRUE,TRUE,TRUE,TRUE))
plot_eem(eemlist[[6]])

#only remove only rayleigh lines
eemlist <- remove_scattering(example_eems, type=c(FALSE,FALSE,TRUE,TRUE))
plot_eem(eemlist[[6]])

#change the width of the lines
eemlist <- remove_scattering(example_eems, width=c(16,3,100,40))
plot_eem(eemlist[[6]])
}
\references{
Gilchrist, J. R., & Reynolds, D. M. (2014). Optical Spectroscopy Instrumentation Design, Quality Assurance, and Control: Bench-Top Fluorimetry. In A. Baker, D. M. Reynolds, J. Lead, P. G. Coble, & R. G. M. Spencer (Eds.), Aquatic Organic Matter Fluorescence (pp. 147-189).
Cambridge: Cambridge University Press. \url{https://doi.org/10.1017/CBO9781139045452.009}
}
