% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{reset_user_config}
\alias{reset_user_config}
\title{Reset all eemanalyzeR settings in the user configuration file to package defaults}
\usage{
reset_user_config()
}
\value{
Invisibly returns the reset default configuration settings as a named list.
}
\description{
This allows the user to overwrite the data processing settings in the user configuration file back to
the default configuration of the eemanalyzeR package. These defaults are documented in data.R under "default_config".
This function is provided in case the user has a malformed configuration file or wants to revert back to default processing
settings after experimenting with modifying the settings using `edit_user_config`.
}
\examples{
reset_user_config()
load_user_config()
}
