% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-eems_draft.R
\name{run_eems}
\alias{run_eems}
\title{Takes raw EEMs and absorbance data from Aqualog, returns cleaned, processed data}
\usage{
run_eems(prjpath, output_dir, filename, interactive = TRUE, ...)
}
\arguments{
\item{prjpath}{path to folder with raw samples}

\item{output_dir}{TODO}

\item{filename}{TODO}

\item{interactive}{TODO}

\item{...}{}
}
\value{
saves processed EEMs as .csv and .rds files, absorbance as .csv and .rds, and metadata as .csv
creates plots for each sample and calculates indices.
}
\description{
Takes raw EEMs and absorbance data from Aqualog, returns cleaned, processed data
}
\details{
The following steps are chosen by default:

 -The file directory is created, raw files are zipped

 -A process file is created, The instrument blank is subtracted from the
 samples, raman scattering is removed and interpolated, rayleigh scattering
 is removed but not interpolated, widths are determined automatically,
 samples are corrected for inner filter effects, raman normalized, and
 normalized for DOC, excitation is clipped from 247 to 450 nm, emission
 is clipped from 247 to 550 nm

 -Sample are plotted individually and a summary plot using the parula
 color palette with descriptions for the plot titles, the peaks are not
 annotated on the plot. Plots are created for DOC normalized data (_DOC),
 and for non normalized data.

 -Fluorescence indices are reported for data that has and hasn't been DOC
 normalized, sample are reported in rows.

 If you want to change these defaults add the appropriate arguments into the
 function to change the defaults for the processing functions.
}
