% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-eems.R
\name{run_eems}
\alias{run_eems}
\title{Load, clean, and process EEMs and absorbance data}
\usage{
run_eems(
  input_dir,
  output_dir = NA,
  filename = NA,
  interactive = TRUE,
  blanklist = NULL,
  qaqc_checks = TRUE,
  ...
)
}
\arguments{
\item{input_dir}{Path to folder containing raw EEMs and/or absorbance files.}

\item{output_dir}{Path to save the processed data.}

\item{filename}{A character string, used for file names.}

\item{interactive}{Logical, used to determine if user input should be used.}

\item{blanklist}{\code{eemlist} of blank files to subtract from samples. A
utomatically uses instrument blanks if not provided.}

\item{qaqc_checks}{Logical, should user QA/QC files be used to check data?}

\item{...}{additional arguments used to make one time modifications to processing
arguments and plotting (\link{plot.eemlist}). List of optional arguments
to modify within the processing are located in \link{default_config}. See details
for more info on modifying processing configuration.}
}
\value{
The following files are saved to the output directory specified.
The list contains:
\itemize{
\item \strong{eemlist:} the \code{eemlist}
\item \strong{abslist:} the \code{abslist}
\item \strong{readme:} a character vector containing information about the data processing steps
\item \strong{meta:} the metadata associated with the samples, may be \code{NULL} if not provided
\item \strong{indices:} a list of EEMs and absorbance indices, may be \code{NULL} if not provided
\item \strong{eem_plot:} a list of EEMs plots, may be \code{NULL} if not provided
\item \strong{abs_plot:} a ggplot2 object of the absorbance data, may be \code{NULL} if not provided
}
}
\description{
The main high-level function of the \link{eemanalyzeR} package.
It is designed for users who want a fast, automated workflow without needing
to write complex R code or string together a bunch of processing functions.
By pointing the function to a folder of raw EEMs and absorbance data,
the function performs the full processing pipeline from start to finish.
}
\details{
The only requirement to use this wrapper function to process EEMs
and absorbance data is an \strong{input directory} containing the raw data files and
a metadata spreadsheet. However, there are many optional arguments for the
user to modify the EEMs and absorbance processing. The names and default values
for these arguments can be found in the documentation for the \link{default_config}.

There are four ways for user to use and modify the processing configuration defaults.
\itemize{
\item \strong{Option 1:} User doesn't change anything, the package defaults are used by \code{run_eems}.
\item \strong{Option 2:} User creates a file (stored on their computer) that has processing defaults that
eemanalyzeR pulls from at load time. This is created using the \code{\link[=edit_user_config]{edit_user_config()}} function.
\item \strong{Option 3:} User modifies the defaults \emph{BEFORE} using the \code{run_eems} function using \code{\link[=modify_config]{modify_config()}} function.
This modifies the settings for the R session and will be applied to any data processing that occurs until
the package is reloaded or the R session is restarted. After the package is reloaded the defaults revert
back to the package defaults or (if they exist) user defaults.
\item \strong{Option 4:} User supplies arguments to \code{run_eems} function that modify processing \emph{ONLY} during that run. These
configuration options will not persist across multiple tries of \code{run_eems} and must be specified each time,
but the argument values will be reported in the readme file.
}
}
