% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-eems_draft.R
\name{run_eems}
\alias{run_eems}
\title{Load, clean, and process EEMs and absorbance data}
\usage{
run_eems(
  input_dir,
  output_dir,
  filename,
  interactive = TRUE,
  blanklist = NULL,
  ...
)
}
\arguments{
\item{input_dir}{Path to folder containing raw EEMs and/or absorbance files.}

\item{output_dir}{Path to save the processed data.}

\item{filename}{A character string, used for file names.}

\item{interactive}{Logical, used to determine if user input should be used.}

\item{blanklist}{eemslist of blank files to subtract from samples. Automatically uses instrument blanks if not provided}

\item{...}{additional arguments used to make one time modifications to processing arguments. See #TODO add text here}
}
\value{
Processed data files are saved to the specified \code{output_dir}. See
\link{export_data} for more details on exports. #TODO: also return outputs as object like export-data?
}
\description{
The main high-level function of the \link{eemanalyzeR} package.
It is designed for users who want a fast, automated workflow without needing
to write complex R code or string together a bunch of processing functions.
By pointing the function to a folder of raw EEMs or absorbance data,
specifying an output directory, and providing a file prefix for naming
outputs, the function performs the full processing pipeline from start to finish.
}
\details{
#TODO: instructions on the ways to modify the parameters
}
