% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-eems_draft.R
\name{run_eems}
\alias{run_eems}
\title{Load, clean, and process EEMs and absorbance data}
\usage{
run_eems(
  input_dir,
  output_dir,
  filename,
  interactive = TRUE,
  blanklist = NULL,
  ...
)
}
\arguments{
\item{input_dir}{Path to folder containing raw EEMs and/or absorbance files.}

\item{output_dir}{Path to save the processed data.}

\item{filename}{A character string, used for file names.}

\item{interactive}{Logical, used to determine if user input should be used.}

\item{blanklist}{eemslist of blank files to subract from samples. Automatically detects instrument blanks if not provided}

\item{...}{additional arguments used to make one time modifications to processing arguments. See}
}
\value{
A list of the processed data which is also saved to the output directory specified.
The list contains:
\itemize{
\item \strong{eemlist:} the \code{eemlist}
\item \strong{abslist:} the \code{abslist}
\item \strong{readme:} a character vector containing information about the data processing steps
\item \strong{meta:} the metadata associated with the samples, may be \code{NULL} if not provided
\item \strong{indices:} a list of EEMs and absorbance indices, may be \code{NULL} if not provided
\item \strong{eem_plot:} a list of EEMs plots, may be \code{NULL} if not provided
\item \strong{abs_plot:} a ggplot2 object of the absorbance data, may be \code{NULL} if not provided
}
}
\description{
The main high-level function of the \link{eemanalyzeR} package.
It is designed for users who want a fast, automated workflow without needing
to write complex R code or string together a bunch of processing functions.
By pointing the function to a folder of raw EEMs or absorbance data,
specifying an output directory, and providing a file prefix for naming
outputs, the function performs the full processing pipeline from start to finish.
}
\details{
#TODO: instructions on the ways to modify the parameters
}
