% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-validation-models.R
\name{save_absorbance_model}
\alias{save_absorbance_model}
\title{Create a long term absorbance or EEM model (for tea stds and blanks)}
\usage{
save_absorbance_model(
  path_to_abs,
  save_path = "data",
  sd_multiplier = 3,
  overwrite = FALSE
)
}
\arguments{
\item{path_to_abs}{location of absorbance data}

\item{save_path}{location to save absorbance model}

\item{sd_multiplier}{accetable standard deviation in data}

\item{overwrite}{overwrite exisiting model?}
}
\value{
a absorbance model of averaged absorbance
}
\description{
Create a long term absorbance or EEM model (for tea stds and blanks)
}
