% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.eemlist}
\alias{[.eemlist}
\title{Subsetting using `[` for eemlist}
\usage{
\method{`[`}{eemlist}(eemlist, i)
}
\arguments{
\item{eemlist}{the eemlist to subset}

\item{i}{the index for subsetting}
}
\description{
Subsetting using `[` for eemlist
}
