% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-samples.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset eemlist or abslist using components}
\usage{
subset_samples(
  x,
  info,
  sample,
  keep = FALSE,
  ignore_case = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{an object of class \code{eemlist} or \code{abslist}}

\item{info}{the name of the component within the \code{eem} or \code{abs} to extract.
see \link[eemR]{eem} for base \code{eem} component names and \link[eemanalyzeR]{abs_read} for base \code{abs} names.
see \link[eemanalyzeR]{add_metadata} for extended \code{eem} component names.}

\item{sample}{a vector of the names or other info to use to select EEM's from \code{eemlist} or absorbance from \code{abslist}}

\item{keep}{logical. If TRUE, the specified sample will be returned. If FALSE, they will be removed.}

\item{ignore_case}{logical, should sample name case should be ignored (TRUE) or not (FALSE). Default is FALSE.}

\item{verbose}{logical determining if removed/extracted eems should be printed on screen.}
}
\value{
An object of class \code{eemlist} or \code{abslist} with the samples excluded or selected based on the 'sample' vector.
}
\description{
Helper function that build upon \link[eemR]{eem_extract} function. Used to select or remove samples based on the info extracted by
\link[eemanalyzeR]{get_sample_info} allowing selection of samples based on components besides just the sample name. This is usually
used in concert with `get_sample_info`
}
\examples{
 #subset by name
names <- get_sample_info(example_eems, "sample")
eem_subset <- subset_samples(example_eems, "sample", names[1]) #default is to remove
#but use keep=T to keep instead
eem_subset <- subset_samples(example_eems,"sample", names[1], keep=TRUE)

#subset by file_name
eemlist <- add_metadata(metadata, example_eems)
names <- get_sample_info(eemlist, "meta_name")
eem_subset <- subset_samples(eemlist, "meta_name", names[1]) #default is to remove

}
