% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract-blank.R
\name{subtract_blank}
\alias{subtract_blank}
\title{Perform blank subtraction on EEM samples}
\usage{
subtract_blank(eem)
}
\arguments{
\item{eem}{an \code{eem} or \code{eemlist} that has been augmented with metadata and blanks. See details for more info.}
}
\value{
an object of class \code{eem} or \code{eemlist} where the sample data (\code{x})
has been subtracted by the blank data (\code{blk_x})
}
\description{
Perform blank subtraction on EEM samples
}
\details{
In order to subtract the blank EEMs data from the sample data, you must first 'attach' the blank data to the \code{eem}
by using the function \link[eemanalyzeR]{add_blanks}. Note that in order to run \link[eemanalyzeR]{add_blanks} you
first need to add the metadata via \link[eemanalyzeR]{add_metadata}.

The processing steps are tracked in two ways:
\itemize{
 \item{processing is tracked for each sample via an attribute attached to each \code{eem}
object. You can visualize this using the function \link[base]{attributes}\code{(eem)}}
\item{overall processing steps are also tracked and output as a readme.txt file in the processed data folder} ## TO DO: update when this is finalized
}
}
\examples{
eem <- add_metadata(metadata,example_eems)
eem <- add_blanks(eem)
eem_sub <- subtract_blank(eem[[1]])
eemlist_sub <- subtract_blank(eem)
}
