% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{user_config}
\alias{user_config}
\alias{edit_user_config}
\alias{load_user_config}
\title{Set up and apply user defaults for data processing}
\usage{
edit_user_config(..., interactive = TRUE)

load_user_config(user_config_file = .user_config_path(), env = .pkgenv)
}
\arguments{
\item{...}{potential paramater names and values to apply to the user config. Only if interactive = FALSE}

\item{interactive}{defaults to TRUE, which will open the user config file for manual editing.
If FALSE, it attempts to apply the values provided in \code{...}}

\item{user_config_file}{The path to the file that stores the user config. Defaults to the yaml file in the default user data directory.}

\item{env}{the environment name to write to}
}
\value{
\itemize{
\item \strong{edit_user_config} opens up \code{user-config.yaml}.
\item \strong{load_user_config} will apply the user defaults from \code{user-config.yaml}
to the package environment
}
}
\description{
Reads the user generated YAML file (\code{user-config.yaml}) which stores the
user specified values for the arguments in \code{\link[=run_eems]{run_eems()}} and applies them
to the package environment (\code{.pkgenv}). This allows the user to specify
processing parameters that are maintained across R sessions.
}
\details{
The defaults are stored in a YAML configuration file on the user data directory.
This function will open up the file so the text can be edited. To save new defaults simply
edit the file and save it. The arguments in this file will overwrite the defaults
set in the package.
}
\examples{

load_user_config()
}
