% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-config.R
\name{edit_user_config}
\alias{edit_user_config}
\alias{load_user_config}
\title{Set up and apply user defaults for data processing}
\usage{
edit_user_config()

load_user_config(
  config_path = rappdirs::user_data_dir("eemanalyzeR"),
  env = .pkgenv
)
}
\arguments{
\item{config_path}{path the YAML file with user default values}

\item{env}{the environment name to write to}
}
\value{
\itemize{
\item \strong{edit_user_config} opens up \code{user-config.yaml}.
\item \strong{load_user_config} will apply the user defaults from \code{user-config.yaml}
to the package environment
}
}
\description{
Reads the user generated YAML file (\code{user-config.yaml}) which stores the
user specified values for the arguments in \code{\link[=run_eems]{run_eems()}} and applies them
to the package environment (\code{.pkgenv}). This allows the user to specify
processing parameters that are maintained across R sessions.
}
\details{
The defaults are stored in a YAML configuration file on the user data directory.
This function will open up the file so the text can be edited. To save new defaults simply
edit the file and save it. The arguments in this file will overwrite the defaults
set in the package.
}
\examples{
edit_user_config()

load_user_config()
}
