% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs-indices.R
\name{usgs_indices}
\alias{usgs_indices}
\title{USGS methods for fluorescence and absorbance indices}
\source{
Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018).
Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence from dissolved organic matter
(USGS Numbered Series No. 2018–1096). Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence
from dissolved organic matter (Vol. 2018–1096). Reston, VA: U.S. Geological Survey. \url{https://doi.org/10.3133/ofr20181096}
}
\usage{
usgs_indices(eemlist, abslist, cuvle = 1)
}
\arguments{
\item{eemlist}{an \code{eemlist} object containing EEM's data. See details for more info.}

\item{abslist}{an \code{abslist} object containing absorbance data.}

\item{cuvle}{cuvette (path) length in cm}
}
\value{
a list with two objects:
\itemize{
  \item eem_index: a data.frame of all the fluorescence indices
  \item abs_index: a data.frame of all the absorbance indices
}
Each data.frame will have four columns:
\itemize{
 \item sample_name: the name of the sample
 \item meta_name: the name of the sample in the metadata if metadata has been added, otherwise the sample name again
 \item index: the name of the index being reported, the index name is a human readable index name followed by the USGS parameter code.
 \item value: the value of the index
}
}
\description{
Calculates commonly used absorbance and fluorescence optical indices from eemlist and abslist.
Reports indices commonly used by the U.S. Geological Survey.
For detailed descriptions and references for indices, Hansen et. al. 2018 (Table 1 and 8).
}
\note{
If absorbance is not at a 1 nanometer interval, absorbance will be interpolated using
\link[zoo]{na.approx} which fills in missing values
using linear interpolation.

If EEMs data is not a 1 nanometer interval, fluorescence will be interpolated
using \link[pracma]{interp2}.
}
\examples{
abslist <- add_metadata(metadata, example_absorbance)
eemlist <- add_metadata(metadata, example_eems)
indices <- usgs_indices(eemlist, abslist)
}
