% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-blanks.R
\name{validate_blanks}
\alias{validate_blanks}
\title{Validate the blanks}
\usage{
validate_blanks(blanklist, plotting_info = NULL, blank_model = NULL)
}
\arguments{
\item{blanklist}{an \code{eemlist} containing blank samples}

\item{plotting_info}{TODO - plotting arguments for ggeem (not implemented yet)}

\item{blank_model}{TODO - model of blank noise calculated from multiple runs averaged
together for automated validation.}
}
\value{
TRUE if blanks meet validation standards, FALSE if not
}
\description{
Will run automated validation checks and plot blanks for manual validation.
}
\details{
If validate_blanks run in interactive console (most RStudio sessions), this will plot the blanks
using ggeem. If run non-interactively (such as in a batch processing or in package tests),
nothing will be plotted, but automatic checks will take place.
}
\examples{
continue <- validate_blanks(eem_get_blank(example_eems))
}
