% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-check-std.R
\name{validate_check_std}
\alias{validate_check_std}
\title{Visually Validate the Check Standard}
\usage{
validate_check_std(abslist, std_dir = .qaqc_dir(), tolerance = 0.2)
}
\arguments{
\item{abslist}{an object of class \code{abslist}}

\item{std_dir}{file path to the QAQC files generated with \link[eemanalyzeR]{create_mdl} and \link[eemanalyzeR]{create_tea_std},
default is a user-specific data directory (\link[rappdirs]{user_data_dir})}

\item{tolerance}{what is the maximum percentage the tea standard can vary from the long-term values without being flagged?}
}
\value{
a \code{ggplot} object with the absorbance of the check standards plotted with the long-term tea standard (dashed line), and the
tolerance thresholds (gray ribbon).
}
\description{
Visually inspect plots of the tea check standards compared to the \link[=create_tea_std]{long-term standard}.
}
\examples{
abslist <- add_metadata(metadata, example_abs)
validate_check_std(abslist, system.file("extdata", package = "eemanalyzeR"))
}
